% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.r
\name{acceptance}
\alias{acceptance}
\title{Get acceptance rates from BayesAss output.}
\usage{
acceptance(x)
}
\arguments{
\item{x}{An object returned by BayesAss}
}
\description{
Get acceptance rates from BayesAss output.
}

