% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{add_gate_to_plot}
\alias{add_gate_to_plot}
\title{Add a gate layer to plot}
\usage{
add_gate_to_plot(p, gate)
}
\arguments{
\item{p}{a plot}

\item{gate}{a gate object}
}
\description{
Add a gate layer to plot
}
