% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aDist_mtx.R
\name{aDist_mtx}
\alias{aDist_mtx}
\title{Calculates the Aitchison (compositional) distance between samples.}
\usage{
aDist_mtx(table)
}
\arguments{
\item{table}{of counts in each bin for each donor (row). Rows must sum to one.}
}
\value{
a distance matrix between donors.
}
\description{
Calculates the Aitchison (compositional) distance between samples.
}

