% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow-def.R
\name{as.flow_def}
\alias{as.flow_def}
\title{Reeading a flow_definition file and checking it.}
\usage{
as.flow_def(x)
}
\arguments{
\item{x}{can be a data.frame or a path for a flow_definition file}
}
\description{
Reeading a flow_definition file and checking it.
}

