% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow-def.R
\name{as.flowdef}
\alias{as.flowdef}
\title{Reeading a flow definition file and checking it.}
\usage{
as.flowdef(x)
}
\arguments{
\item{x}{can be a data.frame or a path for a flow definition file}
}
\description{
Reeading a flow definition file and checking it.
}

