% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flowmat.R
\name{as.flowmat}
\alias{as.flowmat}
\alias{is.flowmat}
\title{flow mat}
\usage{
as.flowmat(x, grp_col, jobname_col, cmd_col, ...)

is.flowmat(x)
}
\arguments{
\item{x}{a data.frame or path to file with flow details in it.}

\item{grp_col}{column used for grouping, default samplename.}

\item{jobname_col}{column specifying jobname, default jobname}

\item{cmd_col}{column specifying commands to run, default cmd}

\item{...}{not used}
}
\description{
as.flowmat(): reads a file and checks for required columns. If x is data.frame checks for required columns.
}

