% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{asinh_trans}
\alias{asinh_trans}
\title{Scaled hyperbolic arc-sine transformation}
\usage{
asinh_trans(..., n = 6, equal.space = FALSE)
}
\arguments{
\item{...}{arguments passed to \code{asinh_transform()}}

\item{n}{desired number of breaks (see \code{flow_trans()})}

\item{equal.space}{whether breaks at equal-spaced intervals (see \code{flow_trans()})}
}
\description{
Scaled hyperbolic arc-sine transformation
}
