% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{asinh_transform}
\alias{asinh_transform}
\title{Scaled hyperbolic arc-sine function}
\usage{
asinh_transform(scale = 5, inverse = FALSE)
}
\arguments{
\item{scale}{scale parameter}

\item{inverse}{use inverse function?}
}
\description{
Scaled hyperbolic arc-sine function
}
