% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_direction}
\alias{assign_direction}
\title{Assign Direction}
\usage{
assign_direction(bearing)
}
\arguments{
\item{bearing}{Numeric vector of bearing angles in degrees.}
}
\value{
Character vector of directions ("NB" for northbound, "SB" for southbound).
}
\description{
Assign cardinal direction based on bearing.
}
\details{
\itemize{
  \item 0-90 degrees or 270-360 degrees: Northbound (NB)
  \item 90-270 degrees: Southbound (SB)
}
}
