% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.r
\name{BA_deviance}
\alias{BA_deviance}
\title{Calculate Bayesian deviance from a BayesAss trace file.}
\usage{
BA_deviance(file, burn, interval)
}
\arguments{
\item{file}{Path to BA3trace.txt}

\item{burn}{Number of burn iterations used}

\item{interval}{Sampling intrval used}
}
\value{
Deviance value
}
\description{
Code modified from supplementary files for: Meirmans, P. G. Nonconvergence in
Bayesian estimation of migration rates. 726–733 (2014).
doi:10.1111/1755-0998.12216
}

