% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{build_flowset_from_df}
\alias{build_flowset_from_df}
\title{Build a flowSet from a data.frame}
\usage{
build_flowset_from_df(
  df,
  origin = NULL,
  chanel_col = setdiff(names(df), c("name", "subset")),
  sample_col = "name"
)
}
\arguments{
\item{df}{data.frame with a column \code{sample_col} containing sample names and
columns with flow variables.}

\item{origin}{flowSet from which to retrieve flowFrames description and parameters slots. 
Ignored if NULL.}

\item{chanel_col}{Names of df's columns to be used as flow variables}

\item{sample_col}{Name of df's column containing sample names}
}
\value{
a flowSet
}
\description{
Build a flowSet from a data.frame
}
\examples{
\dontrun{
utils::data("GvHD", package = "flowCore")
gs <- GatingSet(GvHD)
samples <- flowWorkspace::sampleNames(gs)[1:3]
df <- get_data_gs(gs = gs, sample = samples, subset = "root", Ncells = 1000)
fs <- build_flowset_from_df(df = df, origin = gs@data)
pData(fs)}
}
