% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{calc_arrows}
\alias{calc_arrows}
\title{Title}
\usage{
calc_arrows(x, pdf, verbose = get_opts("verbose"))
}
\arguments{
\item{x}{boxht height of the box, as returned by calc_box}

\item{pdf}{creating pdf of displaying interactively}

\item{verbose}{display verbose messages}
}
\description{
Title
}

