% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{calc_boxdim}
\alias{calc_boxdim}
\title{calc_boxdim}
\usage{
calc_boxdim(x, detailed, pdf)
}
\arguments{
\item{x}{number of jobs}

\item{detailed}{detailed}

\item{pdf}{pdf}
}
\description{
calc_boxdim
}

