% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{calc_fontsize}
\alias{calc_fontsize}
\title{Calculate font size based on the size of the window}
\usage{
calc_fontsize(x)
}
\arguments{
\item{x}{box height}
}
\description{
Calculate font size based on the size of the window
}

