% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_bearing}
\alias{calculate_bearing}
\title{Calculate Bearing}
\usage{
calculate_bearing(segment)
}
\arguments{
\item{segment}{An sf linestring object.}
}
\value{
Numeric vector of bearing angles in degrees (0-360).
}
\description{
Calculate bearing angle of a line segment.
}
\details{
Calculates the angle from north (0 degrees) in clockwise direction.
}
