% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow-def.R
\name{check}
\alias{check}
\title{check consistency of objects
Currently only checks for flowdef}
\usage{
check(x, ...)
}
\arguments{
\item{x}{a flowdef object}

\item{...}{suppled to \code{check.classname} function}
}
\description{
check consistency of objects
Currently only checks for flowdef
}

