% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_args}
\alias{check_args}
\title{checks all the arguments in the parent frame.
None of them should be null.}
\usage{
check_args(ignore, select)
}
\arguments{
\item{ignore}{optionally ignore a few variables for checking.}

\item{select}{optionally only check a few variables of the function.}
}
\description{
This function may be optionally moved to a more generic package.
}

