% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_module.R
\name{CleanUI}
\alias{CleanUI}
\alias{Clean}
\title{Identify and remove anomalies}
\usage{
CleanUI(id)

Clean(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\description{
Identify and remove anomalies

Clean module server function
}
\examples{
\dontrun{
if (interactive()){

  ui <- dashboardPage(
    dashboardHeader(title = "flowAI"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      CleanUI("module")
    )
  )

  server <- function(input, output, session) {
    rval <- reactiveValues()
    observe({
      utils::data("Bcells", package = "flowAI")
      rval$gating_set <- flowWorkspace::GatingSet(Bcells)
    })
    res <- callModule(Clean, "module", rval = rval)
  }
  
  shinyApp(ui, server)

}}
}
