% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_module.R
\name{ClusteringUI}
\alias{ClusteringUI}
\alias{Clustering}
\title{Automatic cluster identification}
\usage{
ClusteringUI(id)

Clustering(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Automatic cluster identification

Clustering module server function
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)
library(flowWorkspace)
library(flowCore)

if (interactive()){

  ui <- dashboardPage(
    dashboardHeader(title = "Clustering"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      ClusteringUI("module")
    )
  )

  server <- function(input, output, session) {

    rval <- reactiveValues()
    observe({
      utils::data("GvHD", package = "flowCore")
      gs <- GatingSet(GvHD)
      rval$gating_set <- gs
    })
    rval <- callModule(Clustering, "module", rval = rval)
  }

  shinyApp(ui, server)

}
}
}
