% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_module.R
\name{clusterUI}
\alias{clusterUI}
\alias{cluster}
\title{clusterUI and cluster}
\usage{
clusterUI(id)

cluster(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
A shiny Module that deals with metadata
}
