% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/flow.R
\name{cmds_to_flow}
\alias{cmds_to_flow}
\title{cmds_to_flow}
\usage{
cmds_to_flow(cmd.list, samplename = "", infomat, q_obj = queue(type =
  "lsf"), flowname = "stage2", execute = FALSE,
  flow_base_path = "/scratch/iacs/flow_pipe/tmp")
}
\arguments{
\item{cmd.list}{}

\item{samplename}{}

\item{infomat}{}

\item{q_obj}{}

\item{flowname}{}

\item{execute}{}

\item{flow_base_path}{}
}
\description{
cmds_to_flow
}

