% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_svd.R
\name{comp_svd}
\alias{comp_svd}
\title{Spectral Map Analysis
Input V must be linear}
\usage{
comp_svd(V)
}
\arguments{
\item{V}{a matrix on the linear scale}
}
\value{
The singular value decomposition of the double centered matrix
}
\description{
Spectral Map Analysis
Input V must be linear
}

