% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compensation_module.R
\name{CompensationUI}
\alias{CompensationUI}
\alias{Compensation}
\title{Visualize, import, export and edit compensation matrices associated with a GatingSet}
\usage{
CompensationUI(id)

Compensation(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Visualize, import, export and edit compensation matrices associated with a GatingSet

Compensation server function
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)

if (interactive()){

  ui <- dashboardPage(
    dashboardHeader(title = "Compensation"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      CompensationUI("module")
    )
  )

  server <- function(input, output, session) {
    rval <- reactiveValues()
    observe({
      utils::data("GvHD", package = "flowCore")
      rval$gating_set <- GatingSet(GvHD)
    })

    res <- callModule(Compensation, "module", rval = rval)
  }

  shinyApp(ui, server)

}
}
}
