% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompFlow.R
\name{CompFlow}
\alias{CompFlow}
\title{Compflow}
\usage{
CompFlow(Data, ICCModel = NULL, ICCType = NULL, ICCUnit = NULL, Density = "No")
}
\arguments{
\item{Subdirectory}{A subdirectory of your working directory, in which you store your dataset CSV files. Required.}

\item{TrainingData}{The name of your training dataset, which is used to generate cutpoints. Required.}

\item{CutPointStatus}{The survival status you wish to use to define your cutpoints. Required.}

\item{CutPointTime}{The time variable you wish to use to define your cutpoints. Required.}

\item{minprop}{The minimum proportion of cases to be included in a group. Default is 0.1 if none is provided.}

\item{Variables}{A list of all variables you wish to be included. Names must match and be identical across datasets. Required.}

\item{Palette}{Use 'Greyscale' if a greyscale plot is wanted. Otherwise, red/blue is the default palette.}
}
\value{
CompFlow returns a scatter plot, a Bland-Altman plot and an accompanying ICC.
}
\description{
Compflow
}
\examples{
CutFlow(Subdirectory = "CutFlowData", TrainingData = "Glasgow", CutPointStatus = "CSS", CutPointTime = "CSS_2017", minprop = 0.1, Variables = c("GD_PercPositiveCellsinHealthyStroma", "GD_PercPositiveCellsinHealthyEpithelium", "GD_PercPositiveCellsinHealthyTissue", "GD_PercPositiveCellsinTumourStroma", "GD_PercPositiveCellsinTumourEpithelium", "GD_PercPositiveCellsinTumourTissue", "CD8_PercPositiveCellsinHealthyStroma", "CD8_PercPositiveCellsinHealthyEpithelium", "CD8_PercPositiveCellsinHealthyTissue", "CD8_PercPositiveCellsinTumourStroma", "CD8_PercPositiveCellsinTumourEpithelium", "CD8_PercPositiveCellsinTumourTissue"))
}
\author{
Hayman, \email{l.hayman.1@research.gla.ac.uk}
This function takes in data for two sets of measurements and produces a comparison.
}
