% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpill_module.R
\name{ComputeSpillUI}
\alias{ComputeSpillUI}
\alias{ComputeSpill}
\title{Compute spillover matrix}
\usage{
ComputeSpillUI(id)

ComputeSpill(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The computed matrix
}
\description{
Compute spillover matrix

ComputeSpill server function
}
