% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{copy_gate}
\alias{copy_gate}
\title{Copy a gate and its children as children of another parent gate(optionnal)}
\usage{
copy_gate(gs, name, parent, copy_children_gates = TRUE)
}
\arguments{
\item{gs}{a GatingSet}

\item{name}{name of the gate to copy}

\item{parent}{name of the parent gate}

\item{copy_children_gates}{logical. Should children gates be copied as well 
(the hierarchy of children gates will be conserved)}
}
\description{
Copy a gate and its children as children of another parent gate(optionnal)
}
