% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins_post_process.R
\name{CQ}
\alias{CQ}
\title{Calculates eigenvalue spectrum at different values of q}
\usage{
CQ(C, seq)
}
\arguments{
\item{C}{A correlation matrix, not yet thresholded}

\item{seq}{A vector of thresholds q}
}
\value{
e The eigenvalues of matrix Q

CQ The ratio of consecutive eigenvalues
}
\description{
Calculates eigenvalue spectrum at different values of q
}

