% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/plot-funcs.R
\name{create_jobs_mat}
\alias{create_jobs_mat}
\title{create_jobs_mat}
\usage{
create_jobs_mat(x)
}
\arguments{
\item{x}{}
}
\description{
create_jobs_mat
}
\examples{
\dontrun{
.create_jobs_mat(x = x)}
}

