% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submit_job.R
\name{create_queue_cmd}
\alias{create_queue_cmd}
\title{create_queue_cmd}
\usage{
create_queue_cmd(jobj, file, index, fobj, ...)
}
\arguments{
\item{file}{path to the output file}
}
\description{
create_queue_cmd
}

