% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutFlow.R
\name{CutFlow}
\alias{CutFlow}
\title{CutFlow}
\usage{
CutFlow(
  Subdirectory,
  TrainingData,
  CutPointStatus,
  CutPointTime,
  minprop = 0.1,
  Variables,
  Greyscale = FALSE
)
}
\arguments{
\item{Subdirectory}{A subdirectory of your working directory, in which you store your dataset CSV files. Required.}

\item{TrainingData}{The name of your training dataset, which is used to generate cutpoints. Required.}

\item{CutPointStatus}{The survival status you wish to use to define your cutpoints. Required.}

\item{CutPointTime}{The time variable you wish to use to define your cutpoints. Required.}

\item{minprop}{The minimum proportion of cases to be included in a group. Default is 0.1 if none is provided.}

\item{Variables}{A list of all variables you wish to be included. Names must match and be identical across datasets. Required.}

\item{Greyscale}{Use 'TRUE' if a greyscale plot is wanted. Otherwise, red/blue is the default palette.}
}
\value{
CutFlow returns a folder with subdirectories, containing your input datasets, the otput of cutpoint generation and coded versions of your datasets.
}
\description{
CutFlow
}
\examples{
CutFlow(Subdirectory = "CutFlowData", TrainingData = "Glasgow", CutPointStatus = "CSS", CutPointTime = "CSS_2017", minprop = 0.1, Greyscale = TRUE, Variables = c( "GD_PercPositiveCellsinHealthyTissue", "GD_PercPositiveCellsinTumourTissue",  "CD8_PercPositiveCellsinHealthyTissue", "CD8_PercPositiveCellsinTumourTissue"))
}
\author{
Hayman, \email{l.hayman.1@research.gla.ac.uk}
This function takes in a dataset(s), produces cut points using the maximally selected rank statistic and recodes your datasets.
}
