% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flow.R
\name{detect_dep_type}
\alias{detect_dep_type}
\title{detect_dep_type}
\usage{
detect_dep_type(x, cmds, prev_job)
}
\arguments{
\item{x}{job object}

\item{cmds}{a string of commands}

\item{prev_job}{previous job name}
}
\description{
detect_dep_type
}

