% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Perform dimensionality reduction}
\usage{
dim_reduction(df, yvar, Ncells = NULL, transformation = NULL,
  y_trans = log10_trans(), perplexity = 50, dims = 2,
  method = "tSNE", check_duplicates = FALSE)
}
\arguments{
\item{df}{a data.frame with only numeric variables.}

\item{yvar}{names of df's variables used to perform dimensionality reduction}

\item{Ncells}{Maximum number of cells without any NA values sampled from df. 
If NULL, all cells without any NA values are used}

\item{transformation}{Named list of \code{trans} objects. 
List names should correspond to variable names.}

\item{y_trans}{default \code{trans} object to be used if \code{transformation} is NULL.}

\item{perplexity}{t-SNE perplexity parameter}

\item{dims}{Number of dimensions}

\item{method}{Name of the method used. Either "tSNE" or "umap"}
}
\value{
a data.frame with additionnal columns : 
"tSNE1" and "tSNE2" for method 'tSNE', "UMAP1" and "UMAP2" for method 'umap'
}
\description{
Perform dimensionality reduction
}
