% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim_reduction_module.R
\name{Dim_reductionUI}
\alias{Dim_reductionUI}
\alias{Dim_reduction}
\title{Perform dimensionality reduction using t-SNE or UMAP}
\usage{
Dim_reductionUI(id)

Dim_reduction(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Perform dimensionality reduction using t-SNE or UMAP

Dim_reduction module  server function
}
