% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.r
\name{dm2ba}
\alias{dm2ba}
\title{Convert divMigrate infiles to BayesAss format.}
\usage{
dm2ba(infile, outfile = NULL)
}
\arguments{
\item{infile}{Path to genepop text file}

\item{outfile}{Path to output file}
}
\description{
This function takes a divMigrate infile and converts it into
a BayesAss infile format. Currently does not support haplotype.
}

