% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plots.R
\name{do_plots}
\alias{do_plots}
\title{makes lineplots}
\usage{
do_plots(bins, dir, markers, threshold, y1, y2)
}
\arguments{
\item{bins}{a list of data matrices, corresponding to bins}

\item{dir}{the directory where the plots should be made}

\item{markers}{vector of strings indicating the markers}

\item{threshold}{the minimum sized bin that should be plotted}

\item{y1, }{y2  vertical plot limits}
}
\value{
a plot in the desired directory (png format)
}
\description{
makes lineplots
}

