% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{ellipse_path}
\alias{ellipse_path}
\title{Return coordinates of points along an ellipse defined by its covariance matrix and its center}
\usage{
ellipse_path(cov, mean, n = 100)
}
\arguments{
\item{cov}{covariance matrix}

\item{mean}{coordinates of the center of the ellipse}

\item{n}{number of points to return along the ellipse}
}
\value{
a data.frame with point cordinates
}
\description{
Return coordinates of points along an ellipse defined by its covariance matrix and its center
}
