% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleModule.R
\name{exampleModuleUI}
\alias{exampleModuleUI}
\alias{exampleModule}
\title{exampleModuleUI and exampleModule}
\usage{
exampleModuleUI(id)

exampleModule(input, output, session, stringsAsFactors = TRUE)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{stringsAsFactors}{logical; Convert strings to factors}
}
\value{
A reactive function returning the imported file
}
\description{
A example shiny module to import a csv file
}
\examples{
\dontrun{
library(shiny)
library(DT)
if (interactive()){
ui <- fluidPage(
  exampleModuleUI("fichier"),
  DTOutput("tableau")
)
server <- function(input, output, session) {
  data <- callModule(exampleModule,"fichier")
  output$tableau <- renderDT({data()})
}
shinyApp(ui, server)
}
}

}
