% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{find_all_parent_gates}
\alias{find_all_parent_gates}
\title{find all parent gates of a given Gate section, recursively}
\usage{
find_all_parent_gates(x)
}
\arguments{
\item{x}{a xml document}
}
\description{
find all parent gates of a given Gate section, recursively
}
