% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{flow}
\alias{flow}
\title{Flow constructor}
\usage{
flow(jobs = list(new("job")), name = "newflow", desc,
  mode = c("scheduler", "trigger", "R"), flow_base_path = "~/flows",
  trigger_path = "", flow_path = "", status = "")
}
\arguments{
\item{jobs}{}

\item{name}{}

\item{desc}{}

\item{mode}{}

\item{flow_base_path}{}

\item{trigger_path}{}

\item{flow_path}{}

\item{status}{}
}
\description{
Flow constructor
}

