% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\docType{data}
\name{opts_flow}
\alias{opts_flow}
\title{options manager for flowr and ngsflows.}
\format{\preformatted{List of 3
 $ get :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 7 14 9 2 21 9 210 212
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x7fe3ff9cf188> 
 $ set :function (..., .dots)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 13 14 15 2 21 9 216 218
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x7fe3ff9cf188> 
 $ load:function (...)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 19 15 21 2 22 9 222 224
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilealias', 'srcfile' <environment: 0x7fe3ff9cf188> 
}}
\usage{
opts_flow
}
\description{
options manager for flowr and ngsflows.
}
\keyword{datasets}

