% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr.R
\name{flowr}
\alias{flowr}
\title{Execute the flowr configuration file}
\usage{
flowr(config)
}
\arguments{
\item{config}{character string containing a path 
to the flowr configuration file}
}
\description{
Execute the flowr configuration file
}
