% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr_import.R
\name{flowr_import}
\alias{flowr_import}
\title{Elegant import of a library}
\usage{
flowr_import(package)
}
\arguments{
\item{package}{character string}
}
\description{
Try to import dependency, if missing, install it
Currently support CRAN and BioConductor repositories.
}
