% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR_server.R
\name{flowR_server}
\alias{flowR_server}
\title{flowR app main server function}
\usage{
flowR_server(session, input, output, modules = NULL)
}
\arguments{
\item{session}{shiny session}

\item{input}{shiny input}

\item{output}{shiny output}

\item{modules}{Names of a shiny modules (server function) to integrate to the app.
The name of the ui function should be made of the name of the module's server function 
with the suffix 'UI'. In addition, the module should take the reactiveValues object 'rval' 
as input and return it as an output.}
}
\description{
flowR app main server function
}
