% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR_server2.R
\name{flowR_server2}
\alias{flowR_server2}
\title{flowR server function}
\usage{
flowR_server2(session, input, output, modules = NULL)
}
\arguments{
\item{session}{shiny session}

\item{input}{shiny input}

\item{output}{shiny output}

\item{module_names}{name of a shiny modules (server function) to integrate to the app}
}
\description{
flowR server function
}
