% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR_ui2.R
\name{flowR_ui2}
\alias{flowR_ui2}
\title{FlowR UI function}
\usage{
flowR_ui2()
}
\arguments{
\item{user_module_name}{name of a user shiny module (ui function) to integrate to the app.
The name of the ui function should be made of the parameter 'user_module_name' with the suffix 'UI'.}
}
\description{
FlowR UI function
}
