% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsets_module.R
\name{flowsetsUI}
\alias{flowsetsUI}
\alias{flowsets}
\title{flowsetsUI and flowsets}
\usage{
flowsetsUI(id)

flowsets(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactive values object \code{rval}
}
\description{
A shiny module dealing with flow-sets and their hierarchy
}
