% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules_module.R
\name{format_info}
\alias{format_info}
\title{Suppresses the function name from its description as 
returned by 'get_package_functions_info()'}
\usage{
format_info(info)
}
\arguments{
\item{info}{a character string}
}
\value{
a character string
}
\description{
Suppresses the function name from its description as 
returned by 'get_package_functions_info()'
}
