% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{format_plot}
\alias{format_plot}
\title{Format a ggplot object}
\usage{
format_plot(p, options = list())
}
\arguments{
\item{p}{a ggplot object}

\item{options}{list of plot format options. Names of options include:
xlim : x-axis range
ylim : y-axis range
transformation : named list of trans objects 
default_trans : default trans object (set to 'identity_trans()' by default). 
Used only if 'transformation' is not an element of 'options'.
axis_labels : named list with axis labels (each element should be named after a plot variable)
color_var_name : name to display for color variable
facet_var : names of the variables used for facetting plots along the x-axis
facet_yvar : names of the variables used for facetting plots along the y-axis
scales : control scaling across facets (passed to 'facet_grid()'), Set to "fixed" by default
option : name of the viridis palette
theme : name of the ggplot theme ("gray" by default)
legend.position : legend position}
}
\value{
a ggplot object
}
\description{
Format a ggplot object
}
