% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSets_module.R
\name{GatingSetsUI}
\alias{GatingSetsUI}
\alias{GatingSets}
\title{Manage and visualize GatingSets and their hierarchy}
\usage{
GatingSetsUI(id)

GatingSets(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
The reactive values object \code{rval} with updated elements :
\describe{
  \item{gating_set_list}{a named list with each element containing :}
  \describe{
    \item{gating_set}{: a GatingSet objects}
    \item{parent}{: the name of its parent GatingSet}
  }
  \item{gating_set_selected}{Name of the selected GatingSet}
}
}
\description{
Manage and visualize GatingSets and their hierarchy

GatingSets module server function
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)
library(flowCore)

if (interactive()){
  
  ui <- dashboardPage(
    dashboardHeader(title = "GatingSets"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      fluidRow(
        column(8, box(width = NULL, GatingSetsUI("module")))
      )
    )
  )
  
  server <- function(input, output, session) {
    
    rval <- reactiveValues()
    
    observe({
      utils::data("GvHD", package = "flowCore")
      gs <- GatingSet(GvHD)
      rval$gating_set_list <- list( GvHD = list(gating_set = gs, parent = NULL),
                                    subset = list(gating_set = gs, parent = "GvHD"))
    })
    
    rval <- callModule(GatingSets, "module", rval = rval)
    
  }
  
  shinyApp(ui, server)
  
}}
}
