% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_all_ancestors}
\alias{get_all_ancestors}
\title{Return all ancestors from a set of nodes in a tree}
\usage{
get_all_ancestors(named_list, names)
}
\arguments{
\item{named_list}{a list representing a tree. It must be named 
according to tree node names and each of its element must have a field 'parent' 
containing the name of its parent node.}

\item{names}{Names of the nodes for which all ancestors must be returned}
}
\description{
Return all ancestors from a set of nodes in a tree
}
