% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_all_descendants}
\alias{get_all_descendants}
\title{Return all descendants from a set of nodes in a tree}
\usage{
get_all_descendants(named_list, names)
}
\arguments{
\item{named_list}{a list representing a tree. It must be named according to tree node names and 
each of its element must have a field 'parent' containing the name of its parent node.}

\item{names}{Names of the nodes for which all descendants must be returned}
}
\description{
Return all descendants from a set of nodes in a tree
}
