% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins_post_process.R
\name{get_bins_lineplot}
\alias{get_bins_lineplot}
\title{splits a datamatrix D into bins with respect to the output of make_bins.}
\usage{
get_bins_lineplot(data_binned, w, D, sample_fac)
}
\arguments{
\item{data_binned}{output of make_bins}

\item{w}{proportion of zero counts (over samples) allowed in a bin}

\item{D}{the matrix to split. Probably compensated, so you can see what the markers are doing}

\item{sample_fac}{factor indicating which sample type each donor originated from, length(sample_fac) == nrow(D)}
}
\value{
bins data matrices for each bin, combined over all donors

binnames The bin names
}
\description{
splits a datamatrix D into bins with respect to the output of make_bins.
}

