% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_cluster}
\alias{get_cluster}
\title{Identify clusters}
\usage{
get_cluster(
  df,
  yvar,
  transformation = NULL,
  y_trans = identity_trans(),
  dc = 3,
  alpha = 0.001,
  k = 100,
  k_meta = 8,
  scale = FALSE,
  method = "FlowSOM"
)
}
\arguments{
\item{df}{a data.frame with only numeric variables.}

\item{yvar}{names of df's variables used to find clusters}

\item{transformation}{Named list of \code{trans} objects. 
List names should correspond to variable names.}

\item{y_trans}{default \code{trans} object to be used if \code{transformation} is NULL.}

\item{dc}{ClusterX dc parameter}

\item{alpha}{ClusterX alpha parameter}

\item{k}{integer; number of nearest neighbours (passed to \code{Rphenograph()})}

\item{k_meta}{Maximum number of clusters to try out 
(passed to \code{FlowSOM::MetaClustering()})}

\item{scale}{logical; Scale values before building SOM (for method 'FlowSOM' only)}

\item{method}{Name of the method used. Either "FlowSOM", "ClusterX", "Rphenograph".}
}
\value{
a data.frame with the additionnal column "cluster"
}
\description{
Identify clusters
}
