% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_data_gs}
\alias{get_data_gs}
\title{Return data from a GatingSet}
\usage{
get_data_gs(
  gs,
  sample = NULL,
  subset = NULL,
  Ncells = NULL,
  spill = NULL,
  return_comp_data = TRUE,
  updateProgress = NULL
)
}
\arguments{
\item{gs}{a GatingSet}

\item{sample}{Names of samples from the GatingSet 
(as returned by \code{pData(gs)$name})}

\item{subset}{Names of subsets from the GatingSet 
(as returned by \code{flowWorkspace::gs_get_pop_paths(gs)})}

\item{Ncells}{number of cells to sample from the GatingSet}

\item{spill}{spillover matrix. If NULL, uncompensated data is used for gating. 
Uncompensated data is returned if parameter 'return_comp_data' is TRUE.}

\item{return_comp_data}{logical. Should compensated data be returned ?}

\item{updateProgress}{function used in shiny to update a progress bar}
}
\value{
a data.frame
}
\description{
Return data from a GatingSet
}
