% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_flow_status}
\alias{flow:::status}
\alias{get_flow_status}
\title{get_flow_status}
\usage{
get_flow_status(x, cores = 6, out_format = "rst")
}
\arguments{
\item{x}{path to the flow; may be without the uuid}

\item{cores}{}
}
\description{
get_flow_status
}
\examples{
\dontrun{
get_flow_status(x = x, cores = 6)}
}

