% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_gate_coordinates}
\alias{get_gate_coordinates}
\title{Return coordinates of flowCore gate.}
\usage{
get_gate_coordinates(gate)
}
\arguments{
\item{gate}{a flowCore gate either from class "polygonGate", "rectangleGate" or 
"ellipsoidGate"}
}
\description{
Return coordinates of flowCore gate.
}
