% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_gates_from_gs}
\alias{get_gates_from_gs}
\title{Build a gating hierarchy from a GatingSet}
\usage{
get_gates_from_gs(gs)
}
\arguments{
\item{gs}{a GatingSet}
}
\value{
a named list representing the gating hierarchy. 
Each element has a field 'gate' with a flowCore filter object 
and a field 'parent' with the name of its parent gate.
}
\description{
Build a gating hierarchy from a GatingSet
}
