% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_gates_from_ws}
\alias{get_gates_from_ws}
\title{Extract all gates from a flowJO workspace}
\usage{
get_gates_from_ws(ws_path, group = NULL)
}
\arguments{
\item{ws_path}{path to the workspace}

\item{group}{Names of the sample groups to be considered}
}
\description{
Extract all gates from a flowJO workspace
}
