% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_gates_from_ws_diva}
\alias{get_gates_from_ws_diva}
\title{Extract all gates from a Diva xml workspace}
\usage{
get_gates_from_ws_diva(ws_path, template = NULL)
}
\arguments{
\item{ws_path}{path to the workspace}

\item{template}{Names of the template to consider}
}
\description{
Extract all gates from a Diva xml workspace
}
