% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geo_dataframe}
\alias{get_geo_dataframe}
\title{Get Geo Data Frame}
\usage{
get_geo_dataframe(df, lat_col, lon_col)
}
\arguments{
\item{df}{A data.table or data.frame.}

\item{lat_col}{Character string specifying the latitude column name.}

\item{lon_col}{Character string specifying the longitude column name.}
}
\value{
An sf object with point geometry.
}
\description{
Convert data.table to sf spatial data frame.
}
