% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_groups_from_ws}
\alias{get_groups_from_ws}
\title{Return the names of all 'groups'
from a FlowJO workspace}
\usage{
get_groups_from_ws(ws_path)
}
\arguments{
\item{ws_path}{path to the workspace}
}
\description{
Return the names of all 'groups'
from a FlowJO workspace
}
