% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.r
\name{get_migrateN_results}
\alias{get_migrateN_results}
\title{Load MIGRATE-N outfile.}
\usage{
get_migrateN_results(path)
}
\arguments{
\item{path}{Path to genepop text file}
}
\value{
a data frame of migration parameters
}
\description{
This function loads MIGRATE-N results as a data frame.
}

