% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules_module.R
\name{get_package_functions_info}
\alias{get_package_functions_info}
\title{get description of package's functions}
\usage{
get_package_functions_info(package_name)
}
\arguments{
\item{package_name}{Name of the package}
}
\value{
A vector with the description of the functions in the package
}
\description{
get description of package's functions
}
