% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_parameters_gs}
\alias{get_parameters_gs}
\title{Get parameters from a GatingSet}
\usage{
get_parameters_gs(gs)
}
\arguments{
\item{gs}{a GatingSet}
}
\value{
a list
}
\description{
Get parameters from a GatingSet
}
