% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Return data used for plotting a GatingSet}
\usage{
get_plot_data(
  gs,
  df = NULL,
  sample,
  subset,
  Ncells = NULL,
  spill = NULL,
  metadata = NULL,
  vartype = NULL
)
}
\arguments{
\item{gs}{a GatingSet}

\item{df}{data.frame with columns \code{name} and \code{subset} 
containing sample and subset names respectively and
columns with plot variables. 
Ignored if \code{NULL}. 
Otherwise supersedes parameters 'gs', 'sample', 'spill', 'metadata'.}

\item{sample}{Names of samples from the GatingSet 
(as returned by \code{pData(gs)$name})}

\item{subset}{Names of subsets from the GatingSet
(as returned by \code{gs_get_pop_paths(gs)})}

\item{Ncells}{number of cells to sample from the GatingSet}

\item{spill}{spillover matrix. If NULL, uncompensated data is returned and used for gating.}

\item{metadata}{a data.frame containing metadata associated to samples.
Must have a column \code{name} used for mapping.}
}
\value{
a data.frame
}
\description{
Return data used for plotting a GatingSet
}
