% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/status-funcs.R
\name{get_resources}
\alias{get_resources}
\title{get_resources}
\usage{
get_resources(x, odir, ...)
}
\arguments{
\item{x}{}

\item{odir}{}
}
\description{
get_resources currenty this only works on LSF
}
\examples{
\dontrun{
get_resources(x = x, odir = ~/tmp)
}
}

