% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zz-parse-resources.R
\name{get_resources_lsf}
\alias{get_resources_lsf}
\title{get_resources_lsf}
\usage{
get_resources_lsf(wd, cores = 4, pattern = "out$", plot = FALSE)
}
\arguments{
\item{wd}{Path to a flow working directory}

\item{cores}{Number of cores to use. [Numeric]}

\item{pattern}{Pattern to use to get lsf stdout files. Defaults to \code{out$}}
}
\description{
get_resources_lsf
}
\examples{
\dontrun{
get_resources_lsf(wd = wd, cores = 4, pattern = out\\$)
}
}
\keyword{internal}

