% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_spillover_matrices_from_ws}
\alias{get_spillover_matrices_from_ws}
\title{Extract spillover matrices from a FlowJO workspace}
\usage{
get_spillover_matrices_from_ws(ws_path)
}
\arguments{
\item{ws_path}{path to the workspace}
}
\description{
Extract spillover matrices from a FlowJO workspace
}
