% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow-det.R, R/status.R
\name{to_flowdet.character}
\alias{get_status}
\alias{get_status.character}
\alias{get_status.data.frame}
\alias{get_status.flow}
\alias{to_flowdet.character}
\title{get_status}
\usage{
\method{to_flowdet}{character}(x)

get_status(x, ...)

\method{get_status}{character}(x, out_format = "markdown", ...)

\method{get_status}{data.frame}(x, ...)

\method{get_status}{flow}(x, out_format = "markdown", ...)
}
\arguments{
\item{x}{flow, flow_det or a path to flow run.}
}
\description{
get a flow_details file from the directory structure. This has less information than the
one generated using a flow object. Lacks jobids etc...

get_status

Uses the trigger column to get the exit status and updates the data.frame.
}

