% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{get_templates_from_ws_diva}
\alias{get_templates_from_ws_diva}
\title{Return the names of all 'templates'
from a Diva xml workspace}
\usage{
get_templates_from_ws_diva(ws_path)
}
\arguments{
\item{ws_path}{path to the workspace}
}
\description{
Return the names of all 'templates'
from a Diva xml workspace
}
