% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_cols}
\alias{get_unique_cols}
\title{Get Unique Columns}
\usage{
get_unique_cols(df)
}
\arguments{
\item{df}{A data.table or data.frame.}
}
\value{
Character vector of column names that have unique values.
}
\description{
Get columns with unique values in a data.table.
}
