% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/generic-funcs.R
\name{get_unique_id}
\alias{get_unique_id}
\title{get_unique_id}
\usage{
get_unique_id(prefix = "id", suffix = "", random_length = 8)
}
\arguments{
\item{random_length}{}

\item{base}{}
}
\description{
get_unique_id
}
\examples{
\dontrun{
get_unique_id(base = id, random_length = 8)}
}

