% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{getPopStatsPlus}
\alias{getPopStatsPlus}
\title{Return statistics for all subsets ans samples in a GatingSet}
\usage{
getPopStatsPlus(gs, spill = NULL, filter = NULL)
}
\arguments{
\item{gs}{a GatingSet}

\item{spill}{spillover matrix. If NULL, uncompensated data is used for gating}

\item{filter}{a filter object applied before computing statistics. Ignored if NULL.}
}
\description{
Return statistics for all subsets ans samples in a GatingSet
}
