% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{gh_get_gate_names}
\alias{gh_get_gate_names}
\title{Get all gate names from a GatingHierarchy}
\usage{
gh_get_gate_names(gh)
}
\arguments{
\item{gh}{a GatingHierarchy}
}
\description{
Get all gate names from a GatingHierarchy
}
