% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalisedRV.R
\name{GRV}
\alias{GRV}
\title{calculates the generalised RV coefficient of two matrices}
\usage{
GRV(GX, GY)
}
\arguments{
\item{GX}{matrix}

\item{GY}{matrix}
}
\value{
RV coefficient
}
\description{
calculates the generalised RV coefficient of two matrices
}

