% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{gs_get_fs_subset}
\alias{gs_get_fs_subset}
\title{Return a flowSet from a given subset in a GatingSet}
\usage{
gs_get_fs_subset(gs, spill = NULL, subset)
}
\arguments{
\item{gs}{a GatingSet}

\item{spill}{spillover matrix. If NULL, uncompensated data is used for gating}

\item{subset}{Name of the subset}
}
\description{
Return a flowSet from a given subset in a GatingSet
}
