% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{haversine_distance}
\alias{haversine_distance}
\title{Calculate Haversine Distance}
\usage{
haversine_distance(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Numeric vector of latitude values for first point(s).}

\item{lon1}{Numeric vector of longitude values for first point(s).}

\item{lat2}{Numeric vector of latitude values for second point(s).}

\item{lon2}{Numeric vector of longitude values for second point(s).}
}
\value{
Numeric vector of distances in meters.
}
\description{
Calculate haversine distance between two points on Earth.
}
\details{
Uses the haversine formula to calculate great-circle distances between points
on a sphere. Earth's radius is assumed to be 6371000 meters.
}
