% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportSpill_module.R
\name{ImportSpillUI}
\alias{ImportSpillUI}
\alias{ImportSpill}
\title{Import spillover matrix}
\usage{
ImportSpillUI(id)

ImportSpill(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The Importd matrix
}
\description{
Import spillover matrix

ImportSpill server function
}
