% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_module.R
\name{ImportUI}
\alias{ImportUI}
\alias{Import}
\title{Import data and build a GatingSet}
\usage{
ImportUI(id)

Import(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object
}
\description{
Import data and build a GatingSet

Import module server function
}
