% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{job}
\alias{job}
\title{job class
#'}
\usage{
job(cmds = "", base_path = "", parent_flow = "", name = "myjob",
  q_obj = new("queue"), submission_type = c("scatter", "serial"),
  status = "", dependency_type = c("none", "gather", "serial", "burst"),
  ...)
}
\arguments{
\item{cmds}{}

\item{base_path}{}

\item{parent_flow}{}

\item{name}{}

\item{q_obj}{}

\item{submission_type}{}

\item{status}{}

\item{dependency_type}{}

\item{...}{}
}
\description{
job class
#'
}
\examples{
q_obj <- queue(type="torque")
j_obj <- job(q_obj=q_obj,cmd="sleep 2",cpu=1)
}

