% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/status-funcs.R
\name{kill_flow}
\alias{kill_flow}
\title{kill_flow}
\usage{
kill_flow(wd, fobj, kill_cmd = "bkill", jobid_col = "job_sub_id")
}
\arguments{
\item{wd}{}

\item{fobj}{}

\item{kill_cmd}{}

\item{jobid_col}{}
}
\description{
kill_flow
}
\examples{
\dontrun{
## example for terminal
flowr kill_flow wd=path_to_flow_directory
}
}

