% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linestrings_to_graph}
\alias{linestrings_to_graph}
\title{Convert Linestring to Graph}
\usage{
linestrings_to_graph(lines, digits = 6, keep.cols = NULL)
}
\arguments{
\item{lines}{An sf data frame of LINESTRING geometries.}

\item{digits}{Numeric rounding applied to coordinates (to ensure that matching points across different linestrings is not impaired by numeric precision issues). Set to \code{NA/Inf/FALSE} to disable.}

\item{keep.cols}{Character vector of column names to keep from the input data frame.}
}
\value{
A data.frame representing the graph with columns:
\itemize{
 \item \code{line} - Line identifier
 \item \code{from} - Starting node ID
 \item \code{FX} - Starting node X-coordinate (longitude)
 \item \code{FY} - Starting node Y-coordinate (latitude)
 \item \code{to} - Ending node ID
 \item \code{TX} - Ending node X-coordinate (longitude)
 \item \code{TY} - Ending node Y-coordinate (latitude)
}
}
\description{
Convert Linestring to Graph
}
\seealso{
\link{simplify_network} \link{flowr-package}
}
