% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.r
\name{load_ba_results}
\alias{load_ba_results}
\title{Load BayesAss outfile.}
\usage{
load_ba_results(path)
}
\arguments{
\item{path}{Path to genepop text file}
}
\value{
a data frame of migration parameters
}
\description{
This function loads BayesAss results as a data frame. Currently it retrieves
only the pairwise population stats, but it may later be extended to retrieve
individual ancestry estimates.
}

