% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\name{load_conf}
\alias{load_conf}
\title{Fetching and loading configuration files.}
\usage{
load_conf(x, check = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{path to a configuration file}

\item{check}{in case of a configuration file, whether to check if files defined in parameters exists..}

\item{...}{Not used}
}
\description{
These are two column tab seperated files (namely name and value).
}
\details{
params ending in: "path$|dir$|exe$" are checked in case of check=TRUE.
}
\seealso{
\link{get_opts}
}

