% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratio_data.R
\name{logratio_data}
\alias{logratio_data}
\title{transforms compositional data for a biplot. First, each row is centered on its geometrical mean, followed by log transformation. This is equivalent to the logratio of every single pair of variables. Then the data is scaled.}
\usage{
logratio_data(data)
}
\arguments{
\item{data}{Matrix of compositions. Each row sums to 1}
}
\value{
data_lr Matrix of logratios, column centered. Suitable for a biplot.
}
\description{
transforms compositional data for a biplot. First, each row is centered on its geometrical mean, followed by log transformation. This is equivalent to the logratio of every single pair of variables. Then the data is scaled.
}

