% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins.R
\name{make_bins}
\alias{make_bins}
\title{bins a dimension-reduced pooled dataset. Can be with PCA or SMA}
\usage{
make_bins(DONORS_reduced, sample_fac, k, l)
}
\arguments{
\item{DONORS_reduced}{pooled dataset decomposed by PCA or SMA. Generally keep all components, because this is chosen within this function}

\item{sample_fac}{index of which sample each event originated from, length(sample_fac) == nrow(D), as given by sample_data}

\item{k}{the number of components to include in analysis}

\item{l}{number of bins along each component}
}
\value{
grid a list of factors, one for each donor, corresponding to the binned events

xtab a table with number of events in each bin, one for each donor.
}
\description{
bins a dimension-reduced pooled dataset. Can be with PCA or SMA
}

