% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biplot.R
\name{make_biplot}
\alias{make_biplot}
\title{make a biplot from a singular value decomposition}
\usage{
make_biplot(s, alpha, dim)
}
\arguments{
\item{s}{A singular value decomposition, returned from svd}

\item{alpha}{either 0 or 1. Commonly alpha = 1, which favours display of the individuals (form biplot). alpha = 0 (covariance biplot) favours display of the variables}

\item{dim}{The number of dimensions to retain}
}
\value{
F Cartesian coordinates of individuals. Columns correspond to components

G Cartesian coordinates of variables. Columns correspond to components. Generally represented as rays emanating from origin.
}
\description{
make a biplot from a singular value decomposition
}

