% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_coexpression.R
\name{make_breaks}
\alias{make_breaks}
\title{Calculates common breaks of all datasets for each channel}
\usage{
make_breaks(X)
}
\arguments{
\item{X}{list of matrices, rows are events, columns are parameters}
}
\value{
list of intervals for each parameter
}
\description{
Calculates common breaks of all datasets for each channel
}

