% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_coexpression.R
\name{make_burt}
\alias{make_burt}
\title{Converts each matrix into a burt matrix, consisting of marker combination and frequency it occurs in that sample}
\usage{
make_burt(X, breaks, markers)
}
\arguments{
\item{X}{list of matrices, rows are events, columns are parameters}

\item{breaks, }{output from make_breaks or custom if desired.}

\item{markers}{vector of character strings indicating marker names, in same order as they appear in the columns of the matrices in X}
}
\value{
list of dataframes, with markers coded as 1 for 'positive' and 0 for 'negative', and additional column 'counts' indicating the proportion of events with that marker combination. Only combinations with non-zero proportions are included. Column 'which' gives a unique index to each sample.
}
\description{
Converts each matrix into a burt matrix, consisting of marker combination and frequency it occurs in that sample
}

