% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_coexpression.R
\name{make_cell_coexp}
\alias{make_cell_coexp}
\title{For every possible pair of markers, find the ratio of the frequency in each sample to the average frequency over all samples.}
\usage{
make_cell_coexp(burt, markers)
}
\arguments{
\item{burt}{output from make_burt}

\item{markers}{vector of character strings, marker names in same order as they appear in columns of burt}
}
\value{
A matrix, where each row is a vector of the ratio of a sample's  frequency to the average  frequency (averaged over all samples). Columns are marker pairs. When a marker combination doesn't appear in the sample, a value of 1 is given.
}
\description{
For every possible pair of markers, find the ratio of the frequency in each sample to the average frequency over all samples.
}

