% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_coexpression.R
\name{make_cross_correlation}
\alias{make_cross_correlation}
\title{Finds the cross-correlation between total frequencies and cell coexpression.}
\usage{
make_cross_correlation(sample_coexp, cell_coexp)
}
\arguments{
\item{sample_coexp}{output from make_sample_coexp}

\item{cell_coexp}{output from make_cell_coexp}
}
\value{
Matrix of cross-correlation values. Rows correspond to total frequencies, columns correspond to cell coexpression frequencies.
}
\description{
Finds the cross-correlation between total frequencies and cell coexpression.
}

