% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid_index.R
\name{make_grid_index}
\alias{make_grid_index}
\title{This takes a combined data matrix (generally dimension has already been reduced) which is gridded by spliting each component into l bins of equal width (defined by minimum and maximum of each component. Thus it is generally a good idea to trim outliers). Helper function for make_grid, not generally used on its own.}
\usage{
make_grid_index(d, l)
}
\arguments{
\item{d}{combined data matrix d, generally dimension reduced}

\item{l}{The number of bins along each parameter}
}
\value{
a factors with which to split d
}
\description{
This takes a combined data matrix (generally dimension has already been reduced) which is gridded by spliting each component into l bins of equal width (defined by minimum and maximum of each component. Thus it is generally a good idea to trim outliers). Helper function for make_grid, not generally used on its own.
}

