% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lineplot_data.R
\name{make_lineplot_data}
\alias{make_lineplot_data}
\title{puts a data matrix (compensated, uncompensated) into a format for lineplots
ncol(X) == length(markers). Hard-coded Convenience function for reshaping.}
\usage{
make_lineplot_data(X, markers)
}
\arguments{
\item{X}{data matrix with observations/events as rows, features/channels as columns}

\item{markers}{The column/channel names, e.g. the biomarkers used}
}
\value{
a data frame for making lineplots/parallel coordinate plots
}
\description{
puts a data matrix (compensated, uncompensated) into a format for lineplots
ncol(X) == length(markers). Hard-coded Convenience function for reshaping.
}

