% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_coexpression.R
\name{make_rsq}
\alias{make_rsq}
\title{Looks at the relationship between correlation of total marker frequency and cell-level coexpression}
\usage{
make_rsq(sample_coexp, cell_coexp, markers)
}
\arguments{
\item{sample_coexp}{output of make_sample_coexp}

\item{cell_coexp}{output of make_cell_coexp}

\item{markers}{vector of character strings, marker names in same order as they appear in columns of burt}
}
\value{
A dataframe. For each pair of markers: R squared value of multiple regression of cell coexpression frequency on both the total frequencies. Correlation between both total frequencies.
}
\description{
Looks at the relationship between correlation of total marker frequency and cell-level coexpression
}

