% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_coexpression.R
\name{make_twoD_marginal}
\alias{make_twoD_marginal}
\title{For every possible pair of markers, find the ratio of the frequency in each sample to expected frequency given no association between the markers.}
\usage{
make_twoD_marginal(burt, markers)
}
\arguments{
\item{burt}{output from make_burt}

\item{markers}{vector of character strings, marker names in same order as they appear in columns of burt}
}
\value{
A matrix, where each row is a vector of the ratio of a sample's  frequency to the expected frequency given no association between the markers. Columns are marker pairs.
}
\description{
For every possible pair of markers, find the ratio of the frequency in each sample to expected frequency given no association between the markers.
}

