% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolism_module.R
\name{metabolism}
\alias{metabolism}
\title{metabolism server function}
\usage{
metabolism(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
metabolism server function
}
