% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata_module.R
\name{MetadataUI}
\alias{MetadataUI}
\alias{Metadata}
\title{Import, export and edit metadata associated with a GatingSet}
\usage{
MetadataUI(id)

Metadata(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Import, export and edit metadata associated with a GatingSet

Metadata module server function
}
\examples{
\dontrun{
# library(shiny)
library(shinydashboard)
library(flowWorkspace)

if (interactive()){

  ui <- dashboardPage(
    dashboardHeader(title = "Metadata"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      MetadataUI("module")
    )
  )

  server <- function(input, output, session) {
    rval <- reactiveValues()
    observe({
      utils::data("GvHD", package = "flowCore")
      rval$gating_set <- GatingSet(GvHD)
    })
    res <- callModule(Metadata, "module", rval = rval)
  }

  shinyApp(ui, server)

}
}
}
