% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.r
\name{mn2ba}
\alias{mn2ba}
\title{Convert MIGRATE-N infiles to BayesAss format.}
\usage{
mn2ba(infile, outfile = NULL)
}
\arguments{
\item{infile}{Path to genepop text file}

\item{outfile}{Path to output file}
}
\description{
This function takes a MIGRATE-N infile (genepop format) and converts it into
a BayesAss infile format.
}

