% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules_module.R
\name{Modules}
\alias{Modules}
\title{Select shiny modules (server)}
\usage{
Modules(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{reactivevalues object with the following elements :
\describe{
  \item{gating_set}{: a GatingSet object}
}}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
A shiny module to search and select other modules
}
