% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins_post_process.R
\name{mtx_thres}
\alias{mtx_thres}
\title{Threshold a correlation matrix such that all values less than q are zero.}
\usage{
mtx_thres(mtx, q)
}
\arguments{
\item{mtx}{A correlation matrix}

\item{q}{The threshold. Values < q are set to zero.}
}
\value{
The thresholded correlation matrix
}
\description{
Threshold a correlation matrix such that all values less than q are zero.
}

