% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization_module.R
\name{NormalizationUI}
\alias{NormalizationUI}
\alias{Normalization}
\title{Normalize data from beads}
\usage{
NormalizationUI(id)

Normalization(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\description{
Normalize data from beads

Normalization module server function
}
