% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\name{get_opts}
\alias{flow_opts}
\alias{get_opts}
\alias{ngsflows_opts}
\alias{print.opts}
\alias{set_opts}
\title{Setting/loading and extracting various options into the environment}
\usage{
get_opts(x)

set_opts(x = list())
}
\arguments{
\item{x}{get_opts: a character vector of names of options to extract.}

\item{x}{a named list}
}
\description{
load_conf:
This function extracts the options set by functions like set_opts/load_conf
}
\examples{
## Prints all the default options set via flowr.conf
get_opts()
get_opts("flow_run_path")
## set _opts
set_opts(list(flow_run_path = "~/mypath"))
}
\seealso{
\link{load_conf}
}

