% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_pretrim.R
\name{outlier_pretrim}
\alias{outlier_pretrim}
\title{trim the data before starting}
\usage{
outlier_pretrim(Data, lim1, lim2)
}
\arguments{
\item{Data}{A list of data matrices}

\item{lim1, }{lim2 The upper and lower limits for trimming, expressed as percentiles. Computed column-wise}
}
\value{
The list of matrices is combined into one single matrix. Then rows are filtered out if at least one value falls outside the bounds set by lim1, lim2. The filted data matrix is returned.
}
\description{
trim the data before starting
}

