% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse-dependency.R
\name{parse_dependency}
\alias{parse_dependency}
\title{Parse dependency
Advanced use, for debugging. Or adding a currenltly un-supported platform.}
\usage{
parse_dependency(x, ...)
}
\arguments{
\item{x}{is a `job` object}
}
\description{
Parse dependency
Advanced use, for debugging. Or adding a currenltly un-supported platform.
}

