% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse-jobids.R
\name{parse_jobids}
\alias{parse_jobids}
\title{parse_jobids}
\usage{
parse_jobids(jobids, platform)
}
\arguments{
\item{jobids}{output from HPCC upon job submission, as a character vector}

\item{platform}{string specifying the platform. This determines how the jobids are parsed}
}
\description{
parse_jobids
}

