% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zz-parse-resources.R
\name{parse_lsf_out}
\alias{parse_lsf_out}
\title{parse LSF output files}
\usage{
parse_lsf_out(x, scale_time = 1/3600, n = 100,
  time_format = get_opts("time_format"))
}
\arguments{
\item{x}{file}

\item{scale_time}{time is usually in seconds, scale of 1/60 shows minutes, 1/3600 shows in hours}

\item{n}{how many lines to read; usually resources details are on top. 100 works well. .Depreciated}

\item{time_format}{format of time in the execution logs. This should match the format in lsf/torque etc. shell script templates.}
}
\description{
parse LSF output files
}

