% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseCompensationDiva}
\alias{parseCompensationDiva}
\title{Return the spillover coefficients associated with a 'parameter' xml_node
from a Diva xml workspace}
\usage{
parseCompensationDiva(parameterNode)
}
\arguments{
\item{parameterNode}{a 'parameter' xml_node from a Diva xml workspace}
}
\description{
Return the spillover coefficients associated with a 'parameter' xml_node
from a Diva xml workspace
}
