% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseGateDiva}
\alias{parseGateDiva}
\title{Return relevant info from a 'gate' xml_node
from a Diva xml workspace}
\usage{
parseGateDiva(gateNode)
}
\arguments{
\item{gateNode}{a 'gate' xml_node from a Diva xml workspace}
}
\description{
Return relevant info from a 'gate' xml_node
from a Diva xml workspace
}
