% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseGroupNodes}
\alias{parseGroupNodes}
\title{Return the name and IDs of samples in a  GroupNode section}
\usage{
parseGroupNodes(x)
}
\arguments{
\item{x}{a xml document}
}
\description{
Return the name and IDs of samples in a  GroupNode section
}
