% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseSpillover}
\alias{parseSpillover}
\title{Return the spillover coefficients associated with a spillover xml_node
from a flowJO wsp}
\usage{
parseSpillover(x)
}
\arguments{
\item{x}{a spillover xml_node from a FlowJO wsp}
}
\description{
Return the spillover coefficients associated with a spillover xml_node
from a flowJO wsp
}
