% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseSpilloverMatrix}
\alias{parseSpilloverMatrix}
\title{Return the spillover matrix associated with a spilloverMatrix xml_node
from a flowJO wsp}
\usage{
parseSpilloverMatrix(x)
}
\arguments{
\item{x}{a spilloverMatrix xml_node from a FlowJO wsp}
}
\description{
Return the spillover matrix associated with a spilloverMatrix xml_node
from a flowJO wsp
}
