% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseSpilloverMatrixDiva}
\alias{parseSpilloverMatrixDiva}
\title{Return the spillover matrix associated with a 'instrument_settings' xml_node
from a Diva xml workspace}
\usage{
parseSpilloverMatrixDiva(settingsNode)
}
\arguments{
\item{settingsNode}{a 'instrument_settings' xml_node from a Diva xml workspace}
}
\description{
Return the spillover matrix associated with a 'instrument_settings' xml_node
from a Diva xml workspace
}
