% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseTemplatesNodesDiva}
\alias{parseTemplatesNodesDiva}
\title{Return the name of 'template' xml_node
from a Diva xml workspace}
\usage{
parseTemplatesNodesDiva(templateNode)
}
\arguments{
\item{templateNode}{a 'template' xml_node from a Diva xml workspace}
}
\description{
Return the name of 'template' xml_node
from a Diva xml workspace
}
