% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_bar}
\alias{plot_bar}
\title{Generates a bar plot}
\usage{
plot_bar(args = list())
}
\arguments{
\item{args}{list of arguments. 
Mandatory argument : a data.frame 'df'
Other arguments can include :
'stat_var' : names of numeric columns to include in the plot
'group_var' : name of the column used to define x axis groups
'color_var' : name of the column used for the aesthetic 'color' in 'ggplot2::geom_point()'
'show.legend' : logical; show legend}
}
\description{
Generates a bar plot
}
