% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_contour}
\alias{plot_contour}
\title{Generates a contour plot}
\usage{
plot_contour(args = list())
}
\arguments{
\item{args}{list of arguments. 
Mandatory arguments : a data.frame 'df', x and y plot variable 'xvar' and 'yvar'.
Other arguments can include :
'color_var' : variable used for the aesthetic 'color'
'group_var' : variable used for the aesthetic 'group'
'show_outliers' : logical; Display all cells in the background
'fill' : fill parameter for contour plot (passed to 'ggplot2::stat_density_2d()')
'bins' : number of grid points in each direction. (passed as parameter 'n' to 'ggplot2::geom_density_2d()')
'alpha' : contour line transparency (between 0 and 1). If 'show_outliers' is TRUE, used to set outliers dot transparency.
'size' : contour line size. If 'show_outliers' is TRUE, used to set outliers dot size.}
}
\description{
Generates a contour plot
}
