% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_dots}
\alias{plot_dots}
\title{Generates a dot plot}
\usage{
plot_dots(args = list())
}
\arguments{
\item{args}{list of arguments. 
Mandatory arguments : a data.frame 'df', x and y plot variable 'xvar' and 'yvar'.
Other arguments can include :
'color_var' : variable used for the aesthetic 'color'
'group_var' : variable used for the aesthetic 'group'
'alpha' : dot transparency (between 0 and 1)
'size' : dot size
'show_label' : logical; add labels for each group (as defined by 'id.vars')
'id.vars' : variable defining groups for which a label should be displayed 
(superseded by 'color_var' and 'group_var')}
}
\description{
Generates a dot plot
}
