% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{plot_flow,flow-method}
\alias{plot_flow}
\alias{plot_flow,flow-method}
\title{plot_flow}
\usage{
\S4method{plot_flow}{flow}(x, detailed = FALSE, pdf = FALSE,
  pdffile = sprintf("\%s.pdf", x@name), ...)
}
\arguments{
\item{x}{Object of class \code{flow}}

\item{detailed}{include some details}

\item{pdf}{create a pdf instead of plotting interactively}

\item{pdffile}{output file name for the pdf file}

\item{...}{experimental}
}
\description{
plot_flow
}
\examples{
plot_flow(x = x, pdf = TRUE)
}

