% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_gate}
\alias{plot_gate}
\title{Plot a gate from a GatingSet}
\usage{
plot_gate(
  gate_name,
  df = NULL,
  gs,
  sample = NULL,
  spill = NULL,
  metadata = NULL,
  plot_type = "contour",
  plot_args = list(),
  options = list()
)
}
\arguments{
\item{gate_name}{name of the gate}

\item{df}{a data.frame with plot data resulting from a call of \code{get_plot_data}. 
Supersedes parameters 'gs', 'sample', 'spill', 'metadata'}

\item{gs}{a GatingSet}

\item{sample}{Set of samples from 'gs'}

\item{spill}{compensation}

\item{metadata}{a data.frame containing metadata associated to samples.
Must have a column \code{name} used for mapping.}

\item{plot_type}{name of the plot type}

\item{plot_args}{list of plot parameters passed to \code{plot_gs()}}

\item{options}{list of plot format options passed to \code{format_plot()}}
}
\description{
Plot a gate from a GatingSet
}
