% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_gh}
\alias{plot_gh}
\title{Plot all gates for a given sample of a gating set}
\usage{
plot_gh(gs, df = NULL, sample = NULL, Ncells = NULL,
  selected_subsets = NULL, spill = NULL, plot_type = "contour",
  plot_args = list(), options = list())
}
\arguments{
\item{gs}{a gating set}

\item{df}{data.frame with columns \code{name} and \code{subset} 
containing sample and subset names respectively and
columns with plot variables. Ignored if \code{NULL}}

\item{sample}{sample names}

\item{selected_subsets}{subset names}

\item{spill}{spillover matrix. If NULL, uncompensated data is used both for gating and plotting.}

\item{...}{parameters passed to \code{plot_gs()}}
}
\value{
a list of ggplot objects
}
\description{
Plot all gates for a given sample of a gating set
}
