% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_gs}
\alias{plot_gs}
\title{Plot data and gates from a gating set}
\usage{
plot_gs(df = NULL, gs = NULL, sample, subset, xvar = NULL,
  yvar = NULL, axis_labels = NULL, color_var = NULL,
  color_var_name = NULL, data_range = NULL, min_value = NULL,
  gate = NULL, polygon_gate = NULL, type = "hexagonal", bins = 100,
  alpha = 0.1, size = 0.1, transformation = NULL,
  default_trans = identity_trans(), spill = NULL, metadata = NULL,
  facet_vars = NULL, yridges_var = "name", norm_density = TRUE,
  smooth = FALSE, ridges = FALSE, bw = 0.1, show.legend = TRUE,
  legend.position = "right", use_log10_count = TRUE,
  theme_name = "theme_gray")
}
\arguments{
\item{df}{data.frame with columns \code{name} and \code{subset} 
containing sample and subset names respectively and
columns with plot variables. Ignored if \code{NULL}}

\item{gs}{a gating set}

\item{sample}{sample names}

\item{subset}{subset names}

\item{xvar}{x variable}

\item{yvar}{y variable}

\item{axis_labels}{Named vector with axis labels. 
Vector names should correspond to variable names}

\item{color_var}{color variable}

\item{data_range}{Named list of data ranges. 
List names should correspond to variable names.}

\item{min_value}{numeric. Minimal value of x and y axis. Ignored if \code{NULL}.}

\item{gate}{A character vector with subset names (i.e elements of \code{getNodes(gs)}) or a list of gates 
(as returned by \code{get_gates_from_gs()}). 
Gates from classes \code{polygonGate}, \code{rectangleGate} and \code{ellipsoidGate} are supported.
If \code{xvar} and \code{yvar} are \code{NULL}, the first gate will be used to set plot variables.}

\item{polygon_gate}{a matrix with columns \code{x} and \code{y} containing path coordinates.}

\item{type}{plot type : either "hexagonal", "histogram", "dots" or "contour"}

\item{bins}{bins parameter}

\item{alpha}{alpha parameter}

\item{size}{size parameter}

\item{transformation}{Named list of \code{trans} objects. 
List names should correspond to variable names.}

\item{default_trans}{default \code{trans} object to be used if \code{transformation} is NULL.}

\item{spill}{spillover matrix. If NULL, uncompensated data is used both for gating and plotting.}

\item{metadata}{data.frame with metadata.}

\item{facet_vars}{variables used to facet plots along the x axis}

\item{yridges_var}{y variables used in \code{geom_density_ridges()}. Used only if \code{ridges} is TRUE.}

\item{norm_density}{normalize maximum density to 1}

\item{smooth}{use \code{geom_density()} instead of \code{geom_histogram()}. Only used if \code{type = "histogram"}.}

\item{ridges}{use \code{geom_density_ridges()}. Only used if \code{type = "histogram"} and \code{smooth = TRUE}.}

\item{bw}{deprecated}

\item{show.legend}{logical. Should plot legend be displayed?}

\item{legend.position}{legend position. "right" by default}

\item{use_log10_count}{logical. Use log-10 transform for color scale if \code{type = "hexagonal"}}

\item{theme_name}{ggplot2 theme name ('theme_gray' by default)}
}
\value{
a ggplot object
}
\description{
Plot data and gates from a gating set
}
