% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Generates a heatmap}
\usage{
plot_heatmap(args = list())
}
\arguments{
\item{args}{list of arguments. 
Mandatory argument : a data.frame or a matrix 'df'
Other arguments can include :
'stat_var' : names of numeric columns to include in the heatmap
'group_var' : name of the column used to order x coordinates
'cluster_y' : logical; cluster y axis variables
'cluster_x' : logical; cluster x axis variables
'show.legend' : logical; show legend
'option' : name of the viridis palette}
}
\description{
Generates a heatmap
}
