% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_hexagonal}
\alias{plot_hexagonal}
\title{Generates a hexagonal heatmap of 2d bin counts (see ggplot2::geom_hex)}
\usage{
plot_hexagonal(args = list())
}
\arguments{
\item{args}{list of arguments. 
Mandatory arguments : a data.frame 'df', x and y plot variables 'xvar' and 'yvar'.
Other arguments can include :
'bins' : numeric, number of bins, passed to 'ggplot2::geom_hex()'
'use_log10_count' : logical, transform bin counts using log10
'option' : name of the viridis palette}
}
\description{
Generates a hexagonal heatmap of 2d bin counts (see ggplot2::geom_hex)
}
