% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Generates an histogram or a density plot}
\usage{
plot_histogram(args = list())
}
\arguments{
\item{args}{list of arguments. 
Mandatory arguments : a data.frame 'df', x plot variable 'xvar'.
Other arguments can include :
'color_var' : variable used for the aesthetic 'color'
'group_var' : variable used for the aesthetic 'group'
'smooth' : logical. Should a density plot be generated?
'ridges': logical. Ignored if 'smooth' is FALSE. Shift different density plots according to 'yridges_var'
'yridges_var' : variable used for the aesthetic 'y' in 'geom_density_ridges'. Ignored if 'ridges' is FALSE. 
'norm_density' : logical. Set maximal y value to 1?
'bins' : number of bins.
(If 'smooth' is TRUE, the inverse of 'bins' is used as the value for the bandwidth parameter 'bw')
'alpha' : transparency (between 0 and 1)}
}
\description{
Generates an histogram or a density plot
}
