% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_pca}
\alias{plot_pca}
\title{Perform a PCA and plot the result}
\usage{
plot_pca(args = list())
}
\arguments{
\item{args}{list of arguments. 
Mandatory argument : a data.frame 'df'
Other arguments can include :
'stat_var' : names of numeric columns used to perform the PCA.
All numeric variables are used by default.
'PCx' : x axis variable ("PC1" by default)
'PCy' : y axis variable ("PC2" by default)
'color_var' : name of the column used to color points
'label_var' : name of the column used to label points
'scale' : logical; scale values by variable before performing the PCA}
}
\description{
Perform a PCA and plot the result
}
