% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_stat}
\alias{plot_stat}
\title{Plot aggregated data from a GatingSet}
\usage{
plot_stat(
  df = NULL,
  gs,
  sample = NULL,
  subset = NULL,
  spill = NULL,
  metadata = NULL,
  transformation = NULL,
  stat_function = "mean",
  y_trans = identity_trans(),
  apply_inverse = TRUE,
  yvar = NULL,
  var_names = NULL,
  scale = FALSE,
  plot_type = "heatmap",
  plot_args = list(),
  options = list()
)
}
\arguments{
\item{df}{a data.frame with plot data resulting from a call of \code{get_plot_data}. 
Supersedes parameters 'gs', 'sample', 'spill'}

\item{gs}{a GatingSet}

\item{sample}{Names of samples from the GatingSet 
(as returned by \code{pData(gs)$name})}

\item{subset}{Names of subsets from the GatingSet 
(as returned by \code{gs_get_pop_paths(gs)})}

\item{spill}{spillover matrix. If NULL, uncompensated data is used for gating and plotting.}

\item{metadata}{a data.frame containing metadata associated to samples.
Must have a column \code{name} used for mapping.}

\item{transformation}{A list of trans objects. 
Each element must be named after a parameter and contain the transfomation to apply for this parameter.
If NULL, the default transformation 'y_trans' is applied.}

\item{stat_function}{Name of the function to be applied on transformed data for each sample and subset.}

\item{y_trans}{Default trans object (set to 'identity_trans()' by default). 
Used only if parameter 'transformation' is NULL.}

\item{apply_inverse}{logical; Apply inverse transformation before returning the data.}

\item{yvar}{Names of the variables for which to compute statistics}

\item{var_names}{Replace 'yvar' with custom names}

\item{scale}{logical; Should data be scaled for each 'yvar' variable?}

\item{plot_type}{name of the plot type}

\item{plot_args}{list of plot parameters passed to the plot function. 
Plot parameters depend on the plot type selected.}

\item{options}{list of plot format options passed to \code{format_plot()}}
}
\value{
a list with a plot and the corresponding plot data
}
\description{
Plot aggregated data from a GatingSet
}
