% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_tile}
\alias{plot_tile}
\title{Generates a tile plot}
\usage{
plot_tile(args = list())
}
\arguments{
\item{args}{list of arguments. 
Mandatory argument : a data.frame 'df'
Other arguments can include :
'stat_var' : names of numeric columns to include in the plot
'group_var' : name of the column used to define x axis groups
'show.legend' : logical; show legend
'option' : name of the viridis palette}
}
\description{
Generates a tile plot
}
