% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_tree}
\alias{plot_tree}
\title{Build a tree graph from a named list}
\usage{
plot_tree(
  gates,
  fontsize = 40,
  rankdir = "LR",
  shape = "ellipse",
  fixedsize = FALSE
)
}
\arguments{
\item{gates}{a named list. Each list element should contain a item 'parent' with 
the name of its parent list element}
}
\description{
Build a tree graph from a named list
}
