% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto_module.R
\name{plotCytoUI}
\alias{plotCytoUI}
\alias{plotCyto}
\title{Build plots from a GatingSet}
\usage{
plotCytoUI(id)

plotCyto(
  input,
  output,
  session,
  rval,
  plot_params = reactiveValues(),
  simple_plot = TRUE,
  auto_update = TRUE,
  show_gates = FALSE,
  polygon_gate = NULL,
  use_ggcyto = FALSE
)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{a reactivevalues object with the following elements :
\describe{
  \item{gating_set}{: a GatingSet object}
  \item{apply_trans}{: logical; apply transformations defined in \code{rval$gating_set}}
  \item{apply_comp}{: logical; apply compensation defined in \code{rval$gating_set}}
}}

\item{plot_params}{reactivevalues object used to initialize plot parameters. 
Amongst others it can contain the following elements (not mandatory):
\describe{
  \item{plot_type}{: type of plot}
  \item{sample}{: names of samples}
  \item{subset}{: names of subsets}
  \item{xvar}{: x-axis variables}
  \item{yvar}{: y-axis variables}
  \item{color_var}{: color aesthetic}
  \item{group_var}{: group aesthetic}
  \item{facet_var}{: variable used to show different plot facets}
  \item{split_var}{: variable used to create multiple plots}
 }}

\item{simple_plot}{logical, disable a number of plot options 
(such as faceting, multiple plots with different x, y or color variables)}

\item{auto_update}{Should plot update be automatic? 
If FALSE, plot update is controlled by an action button.}

\item{show_gates}{Should gates with coordinates matching plot coordinates be displayed}

\item{polygon_gate}{a reactiveValues object with polygon coordinates to be plotted as 
an additionnal layer.
Must contain elements :
\describe{
  \item{x}{vector of x coordinates}
  \item{y}{vector of y coordinates}
}}
}
\value{
A reactivevalues object with the following elements :
\describe{
  \item{plot}{a plot or a list of plots}
  \item{params}{plot parameters}
}
}
\description{
Build plots from a GatingSet

plotCyto module server function
}
