% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGatingHierarchyOutput_module.R
\name{plotGatingHierarchyOutput}
\alias{plotGatingHierarchyOutput}
\alias{plotGatingHierarchy}
\title{plotGatingHierarchyOutput and plotGatingHierarchy}
\usage{
plotGatingHierarchyOutput(id)

plotGatingHierarchy(input, output, session, rval,
  plot_params = reactiveValues())
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
A shiny Module that deals with metadata
}
