% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGatingSet_module.R
\name{plotGatingSet}
\alias{plotGatingSet}
\title{plotGatingSet}
\usage{
plotGatingSet(
  input,
  output,
  session,
  rval,
  plot_params = reactiveValues(),
  simple_plot = TRUE,
  auto_update = TRUE,
  show_gates = FALSE,
  polygon_gate = NULL
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{a reactivevalues object with the following elements :
\describe{
  \item{gating_set}{: a GatingSet object}
}}

\item{plot_params}{reactivevalues object used to initialize plot parameters 
with elements (not mandatory):
\describe{
  \item{samples}{: initially selected samples}
  \item{gate}{: initially selected subsets}
 }}

\item{simple_plot}{logical, disable a number of plot options 
(such as faceting, multiple plots with different x, y or color variables)}

\item{auto_update}{Should plot update be automatic? 
If FALSE, plot update is controlled by an action button.}

\item{show_gates}{Should gates with coordinates matching plot coordinates be displayed}

\item{polygon_gate}{a reactiveValues object with polygon coordinates to be plotted as an additionnal layer.
Must conatain elements :
\describe{
  \item{x}{vector of x coordinates}
  \item{y}{vector of y coordinates}
}}
}
\value{
A reactivevalues object with the following elements :
\describe{
  \item{plot}{a plot or a list of plots}
  \item{params}{plot parameters}
}
}
\description{
A shiny module (server function) to build plots from a GatingSet
}
\details{
plotGatingSet2 server function
}
