% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGatingSetInput_module2.R
\name{plotGatingSetInput}
\alias{plotGatingSetInput}
\alias{plotGatingSet}
\title{plotGatingSetInput and plotGatingSet}
\usage{
plotGatingSetInput(id, simple_plot = TRUE, auto_update = TRUE)

plotGatingSet(input, output, session, rval,
  plot_params = reactiveValues(), simple_plot = TRUE,
  auto_update = TRUE, show_gates = FALSE, polygon_gate = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
A shiny Module that deals with metadata
}
