% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStat_module.R
\name{plotStatInput}
\alias{plotStatInput}
\title{plotStatInput and plotStat}
\usage{
plotStatInput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module to build plots with data aggregation
}
