% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStat_module.R
\name{plotStat}
\alias{plotStat}
\title{plotStat server function}
\usage{
plotStat(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
a list containing the plot, the plot data, and the corresponding plot parameters
}
\description{
plotStat server function
}
