% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting2_module.R
\name{plotting2UI}
\alias{plotting2UI}
\title{plotting2UI and plotting2}
\usage{
plotting2UI(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module to build, display and save plots from a gating set
}
