% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_module.R
\name{PlottingUI}
\alias{PlottingUI}
\title{Build, display and save plots from a gating set}
\usage{
PlottingUI(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
Build, display and save plots from a gating set
}
