% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{plotUI}
\alias{plotUI}
\alias{plot}
\title{plotUI and plot}
\usage{
plotUI(id)

plot(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
A shiny Module that deals with metadata
}
