% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rw-flow-details.R
\name{read_fobj}
\alias{read_fobj}
\title{read flow object given a flow execution folder}
\usage{
read_fobj(x)
}
\arguments{
\item{x}{path to a flow execution folder}
}
\value{
if it finds a fobj, returns that. If not return back the path x
}
\description{
read flow object given a flow execution folder
}

