% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{read_sample_sheet}
\alias{read_sample_sheet}
\title{read_sample_sheet}
\usage{
read_sample_sheet(x, id_column, start_row = 1, sheet = "sample_sheet", ...)
}
\arguments{
\item{x}{}

\item{...}{passed onto \link{read.xlsx}}
}
\description{
read_sample_sheet
}

