% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render-dependency.R
\name{render_dependency}
\alias{render_dependency}
\title{render dependency
Advanced use, for debugging. Or adding a currenltly un-supported platform.}
\usage{
render_dependency(x, ...)
}
\arguments{
\item{x}{is a `job` object}

\item{...}{not used}
}
\description{
render dependency
Advanced use, for debugging. Or adding a currenltly un-supported platform.
}

