% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submit-job.R
\name{render_queue_cmd}
\alias{render_queue_cmd}
\title{render_queue_cmd}
\usage{
render_queue_cmd(jobj, file, index, fobj)
}
\arguments{
\item{jobj}{job object}

\item{file}{path to the output file}

\item{index}{If more than one, which command to focus on. Can be from \code{1:length(cmds)}}

\item{fobj}{flow object}
}
\description{
render_queue_cmd
}

