% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{replace_slots}
\alias{replace_slots}
\title{replace slots in a S4 object}
\usage{
replace_slots(object, ...)
}
\arguments{
\item{object}{a S4 object}

\item{...}{set of slot names to be replaced. This needs to be a named vector}
}
\description{
replace slots in a S4 object
}
\keyword{internal}

