% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logratio_data.R
\name{replace_zero}
\alias{replace_zero}
\title{replaces zero with a small value}
\usage{
replace_zero(X)
}
\arguments{
\item{X}{matrix of compositions}
}
\value{
matrix of compositions, with zero replaced by a small value and rescaled
}
\description{
replaces zero with a small value
}

