% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/rerun-funcs.R
\name{rerun_flow}
\alias{rerun_flow}
\title{rerun_flow}
\usage{
rerun_flow(wd, fobj, execute = TRUE, kill = TRUE)
}
\arguments{
\item{wd}{}

\item{fobj}{}

\item{execute}{}

\item{kill}{logical indicating whether to kill the jobs from old flow}
}
\description{
rerun_flow
}
\examples{
\dontrun{
rerun_flow(wd = wd, fobj = fobj, execute = TRUE, kill = TRUE)
}
}

