% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_col_center.R
\name{row_col_center}
\alias{row_col_center}
\title{pre-processing a data matrix for spectral map analysis}
\usage{
row_col_center(V)
}
\arguments{
\item{V}{A matrix with rows as observations, columns as features}
}
\value{
V The log-transformed matrix is row and column centered, where the row and column means have been weighted by linear column and row sums respectively.
}
\description{
pre-processing a data matrix for spectral map analysis
}

