% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_flowR.R
\name{run_flowR}
\alias{run_flowR}
\title{Run flowR GUI}
\usage{
run_flowR(modules = NULL, shiny.maxRequestSize = 1000)
}
\arguments{
\item{modules}{names of shiny modules (server functions) to include in the GUI}

\item{shiny.maxRequestSize}{file size limit (in MB)}
}
\description{
Run flowR GUI
}
