% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_flowR2.R
\name{run_flowR2}
\alias{run_flowR2}
\title{Run flowR GUI}
\usage{
run_flowR2(modules = NULL, shiny.maxRequestSize = 1000)
}
\arguments{
\item{shiny.maxRequestSize}{file size limit (in MB)}

\item{module_name}{name of shiny module to include in the GUI}
}
\description{
Run flowR GUI
}
