% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{take samples from data. Always use with set.seed for reproducibility}
\usage{
sample_data(Data, D_comp, S)
}
\arguments{
\item{Data}{is a list of uncompensated data matrices}

\item{D_comp}{is a list of compensated data matrices}

\item{S}{is the number of samples to be taken}
}
\value{
samples A list of sample indices applied to each matrix

sample_fac a factor to split combined data matrix into individuals

D combined sampled uncompensated matrix

D_comp combined sampled compensated matrix
}
\description{
take samples from data. Always use with set.seed for reproducibility
}

