% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow-def.R
\name{sample_flow_def}
\alias{sample_flow_def}
\title{Create a skeleton flow definition
A helper function to create a skeleton flow_definition.}
\usage{
sample_flow_def(fl, jobnames)
}
\arguments{
\item{fl}{path to a matrix with commands to run}

\item{jobnames}{names of the jobs in a flow}
}
\description{
Create a skeleton flow definition
A helper function to create a skeleton flow_definition.
}
\details{
flow_tab: as defined by fl is a (minimum) three column matrix with
samplename, jobname, cmd
}

