% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{scale_values}
\alias{scale_values}
\title{scale column of a data frame}
\usage{
scale_values(df, id.vars = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{id.vars}{Names of df's columns that should not be scaled}
}
\value{
a data.frame
}
\description{
scale column of a data frame
}
