% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\name{search_conf}
\alias{search_conf}
\title{Given a name of a configuration file, searches for it in few places}
\usage{
search_conf(x = "flowr.conf", places, verbose = FALSE)
}
\arguments{
\item{x}{name of the file to search for}

\item{places}{places (paths) to look for it. Its best to use the defaults}

\item{verbose}{be chatty?}
}
\description{
Given a name of a configuration file, searches for it in few places
}
\details{
sequence: the later overrides former if parameter name matches
~/                                 ## home dir
~/flowr/conf                       ## flowr default home
MYRLIB/flowr/conf folder           ## flow pipeline folder
}
\examples{
{
search_conf("torque.sh")
}
}

