% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_module.R
\name{selectionInput}
\alias{selectionInput}
\alias{selection}
\title{selectionInput and selection}
\usage{
selectionInput(id, multiple_subset = TRUE)

selection(input, output, session, rval, params = reactiveValues())
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{Global reactivevalues object}

\item{params}{reactivevalues object with initialization values for "samples" and "gate"}
}
\value{
a reactivevalues object with input values
}
\description{
A shiny Module for selecting samples and gates
}
