% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup and initialize some scripts.}
\usage{
setup(bin = "~/bin", flow_base_path = get_opts("flow_base_path"),
  flow_run_path = get_opts("flow_run_path"),
  flow_conf_path = get_opts("flow_conf_path"))
}
\arguments{
\item{bin}{path to bin folder}

\item{flow_base_path}{the root folder for all flowr operations}
}
\description{
Setup and initialize some scripts.
}
\details{
Will add more to this to identify cluster and aid in other things
}

