% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleDisplay_module.R
\name{simpleDisplayUI}
\alias{simpleDisplayUI}
\alias{simpleDisplay}
\title{simpleDisplayUI and simpleDisplay}
\usage{
simpleDisplayUI(id, nrow = 1, size = 400, save = TRUE)

simpleDisplay(
  input,
  output,
  session,
  plot_list,
  params = reactiveValues(use_plotly = FALSE)
)
}
\arguments{
\item{id}{shiny id}

\item{nrow}{number of rows in the layout}

\item{size}{Initial size of a single plot (in pixels)}

\item{save}{logical, add a box with a save button}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{plot_list}{list of plots to display}

\item{params}{a reactivevalues object with an option to display plotly objects}
}
\value{
A list containing the plot displayed and input parameters (these include events 
describing user interaction with the plot)
}
\description{
A shiny Module to display and save plots
}
