% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{split_multi_dep}
\alias{split_multi_dep}
\title{split_multi_dep
Split rows with multiple dependencies}
\usage{
split_multi_dep(x)
}
\arguments{
\item{x}{this is a flow def}
}
\description{
split_multi_dep
Split rows with multiple dependencies
}

