% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayUI_module.R
\name{display}
\alias{display}
\title{display server function}
\usage{
display(input, output, session, rval, module_server_name, ...)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
display server function
}
