% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{status}
\usage{
status(x, cores = 6, out_format = "markdown", get_mem_usage = TRUE)
}
\arguments{
\item{x}{path to the flow; may be without the uuid.}

\item{cores}{number of cores to use}

\item{out_format}{passed onto knitr:::kable. supports: markdown, rst, html...}

\item{get_mem_usage}{under progress, whether to extract mem_usage of jobs}
}
\description{
status
}
\details{
If x is a path with a single flow, it outputs the status of just that.
If the path has more than one flow then this could give a summary of **all** of them.
Instead if x is supplied with paths to more than one flow, then this individually prints status of each.
}
\examples{
\dontrun{
status(x = x, cores = 6)
## an example for running from terminal
flowr status x=path_to_flow_directory cores=6
}
}

