% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submit_flow.R
\name{submit_flow}
\alias{submit_flow}
\title{submit_flow}
\usage{
submit_flow(x, ...)
}
\arguments{
\item{...}{Any additional parameter are passed on to \link{submit_job} function}

\item{fobj}{\code{object} of class \code{flow}.}

\item{uuid}{\code{character} A character string pointing to the folder (unique) where all the logs and other files are processed. This is optional and defaults to:
\code{FLOW_DESCRIPTION_UUID}, and this folder is typically created in \code{~/flows/FLOW_NAME}.
Refer to \code{desc} and \code{name} paramters of \link{flow}.}

\item{execute}{\code{logical} whether or not to submit the jobs}

\item{plot}{\code{logical} whether to make a flow plot (saves it in the flow working directory)}

\item{verbose}{logical.}
}
\description{
submit_flow
}
\examples{
\dontrun{
submit_flow(fobj = fobj, ... = ...)}
}

