% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{submit_job,job,ANY-method}
\alias{submit_job,job,ANY-method}
\title{submit_job}
\usage{
\S4method{submit_job}{job,ANY}(j_obj, execute = FALSE, verbose = TRUE, wd,
  ...)
}
\arguments{
\item{j_obj}{}

\item{...}{}

\item{f_obj}{}
}
\description{
submit_job
}
\examples{
submit_job(j_obj = j_obj, f_obj = f_obj, ... = ...)
}

