% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{submit_job}
\alias{submit_job}
\alias{submit_job,job,flow,job-method}
\alias{submit_job,job,flow-method}
\title{submit_job}
\usage{
submit_job(j_obj, f_obj, ...)

\S4method{submit_job}{job,flow}(j_obj, f_obj, execute = FALSE,
  verbose = FALSE, wd, job_id, ...)
}
\arguments{
\item{j_obj}{job object}

\item{f_obj}{flow object}

\item{...}{not used}

\item{execute}{execute: logical}

\item{verbose}{logical}

\item{wd}{working direcotry}

\item{job_id}{job id}
}
\description{
submit_job
}
\examples{
\dontrun{
submit_job(j_obj = j_obj, f_obj = f_obj, execute = FALSE, verbose = TRUE, wd = wd, job_id = job_id, ... = ...)
}
}

