% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/limit_runs.R
\name{submit_run}
\alias{submit_run}
\title{Submit several flow objects, limit the max running concurrently}
\usage{
submit_run(x, wd, max_processing = 7)
}
\arguments{
\item{x}{a list of flow objects}

\item{wd}{a folder to monitor (flow_run_path)}

\item{max_processing}{max number of flow which may be processed concurrently}
}
\description{
Submit several flow objects, limit the max running concurrently
}

