% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsample_module.R
\name{Subsample}
\alias{Subsample}
\title{Subsample module server function}
\usage{
Subsample(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Subsample module server function
}
