% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rerun.R
\name{subset_fdet}
\alias{subset_fdet}
\title{subset flow details file}
\usage{
subset_fdet(fobj, det, start_from, select, ignore)
}
\arguments{
\item{fobj}{flow object}

\item{det}{flowdet}

\item{start_from}{which job to start from, this is a job name.}

\item{select}{select a subset of jobs to rerun [character vector]}

\item{ignore}{ignore a subset of jobs to rerun [character vector]}
}
\description{
subset flow details file
}

