% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvFlow.R
\name{SurvFlow}
\alias{SurvFlow}
\title{Survflow}
\usage{
SurvFlow(
  Data,
  Variables,
  LegendLabels = NULL,
  Identifier,
  PlotTitles = NULL,
  SurvivalStatus,
  SurvivalTime,
  SurvivalTimeUnit,
  xYearSurvivalVar = 5,
  SurvBase = FALSE
)
}
\arguments{
\item{Data}{A Coded dataset}

\item{Variables}{A list of variables (coded as 0 and 1) for analysis}

\item{LegendLabels}{Optional labels for legend}

\item{Identifier}{Identifier variable for cases}

\item{PlotTitles}{Optional plot titles}

\item{SurvivalStatus}{A status variable (coded as 0 and 1) for survival analysis}

\item{SurvivalTime}{A survival time variable - continuous}

\item{SurvivalTimeUnit}{The unit of time for survival time. Months etc}

\item{xYearSurvivalVar}{The number of years to be used to calculate 'X' years survival. Default = 5}

\item{SurvBase}{Takes TRUE or FALSE. Toggle to activate base survival analysis.}
}
\value{
Takes a coded dataset (use CutFlow) and runs survival analysis for each variable, by each survival outcome, and returns output.
}
\description{
Survflow
}
\examples{
SurvFlow(Glasgow,  Variables = c("GD_PercPositiveCellsinHealthyTissue_Coded", "GD_PercPositiveCellsinTumourTissue_Coded", "CD8_PercPositiveCellsinHealthyTissue_Coded", "CD8_PercPositiveCellsinTumourTissue_Coded"), LegendLabels = c("Low", "High"), Identifier = "TMA_ID", PlotTitles = c("GD Healthy Stroma", "GD Healthy Epithelium", "GD Healthy Tissue", "GD Tumour Stroma", "GD Tumour Epithelium", "GD Tumour Tissue", "CD8 Healthy Stroma", "CD8 Healthy Epithelium", "CD8 Healthy Tissue", "CD8 Tumour Stroma", "CD8 Tumour Epithelium", "CD8 Tumour Tissue"), SurvivalStatus = c("CSS", "OS", "DFS", "RFS"), SurvivalTime = c("CSS_2017", "CSS_2017", "DFSmonths", "DFSmonths"), xYearSurvivalVar = 5, SurvivalTimeUnit = "Months", SurvBase = TRUE)
}
\author{
Hayman, \email{l.hayman.1@research.gla.ac.uk}
This function takes in a dataset(s), produces cut points, recodes your datasets and runs appropriate survival analysis.
}
