% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{test_queue,queue-method}
\alias{test_queue,queue-method}
\title{test_queue}
\usage{
\S4method{test_queue}{queue}(q_obj, verbose = FALSE)
}
\arguments{
\item{q_obj}{}

\item{...}{}
}
\description{
test_queue
}
\examples{
test_queue(q_obj = q_obj, ... = ...)
}

