% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zz-test-queue.R
\name{test_queue}
\alias{test_queue}
\title{test_queue}
\usage{
test_queue(q_obj, verbose = TRUE, ...)
}
\arguments{
\item{q_obj}{queue object}

\item{verbose}{toggle}

\item{...}{These params are passed onto \code{queue}. \code{?queue}, for more information}
}
\description{
This function attempts to test the submission of a job to the queue.
We would first submit one single job, then submit another with a dependency to see if configuration works. This would create a folder in home called 'flows'.
}
\examples{
\dontrun{
test_queue(q_obj = q_obj, ... = ...)}
}

