% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{to_df.status}
\alias{to_df.status}
\title{to_df.status}
\usage{
to_df.status(x)
}
\arguments{
\item{x}{folder with multiple runs}
}
\value{
a data.frame with status per dir.
}
\description{
convert the status to a df. To be called from R
}
