% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flow.r
\name{to_flow.data.frame}
\alias{to_flow.data.frame}
\title{this operates on a single sample basis}
\usage{
\method{to_flow}{data.frame}(x, def, qobj, platform, cpu = 1,
  walltime = "1:00", memory = "1g", flowname, desc, flow_base_path)
}
\description{
this operates on a single sample basis
}
\details{
subset the data.frame by sample and then supply to this function, if you want seperate flow for each sample
}

