% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flow.R
\name{to_flow.list}
\alias{to_flow.list}
\title{to_flow.list}
\usage{
\method{to_flow}{list}(x, def, flowname, flow_run_path, desc, ...)
}
\arguments{
\item{desc}{final flow name}
}
\description{
to_flow.list
}

