% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow-def.R
\name{to_flowdef.character}
\alias{to_flowdef.character}
\title{Create a skeleton flow definition
A helper function to create a skeleton flow definition.}
\usage{
\method{to_flowdef}{character}(x, jobnames, ...)
}
\arguments{
\item{jobnames}{names of the jobs in a flow}

\item{fl}{path to a matrix with commands to run}
}
\description{
Create a skeleton flow definition
A helper function to create a skeleton flow definition.
}
\details{
flow_tab: as defined by fl is a (minimum) three column matrix with
samplename, jobname, cmd
}

