% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow-def.R
\name{to_flowdef}
\alias{to_flowdef}
\alias{to_flowdef.flowmat}
\title{Crate a skeleton flow definition using a flowmat !}
\usage{
to_flowdef(x, ...)

\method{to_flowdef}{flowmat}(x, sub_type = "serial", dep_type = "gather",
  queue = "short", platform = "torque", memory_reserved = "2000",
  cpu_reserved = "1", walltime = "1:00")
}
\arguments{
\item{x}{can a path to a flowmat, flowmat object.}
}
\value{
Returns a flow object. If execute=TRUE, fobj is rich with information about where and how
the flow was executed. It would include details like jobids, path to exact scripts run etc.
To use kill_flow, to kill all the jobs one would need a rich flow object, with job ids present.
}
\description{
Crate a skeleton flow definition using a flowmat !
}

