% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flowmat.R
\name{as.flowmat}
\alias{as.flowmat}
\alias{is.flowmat}
\alias{to_flowmat}
\alias{to_flowmat.data.frame}
\alias{to_flowmat.flow}
\alias{to_flowmat.list}
\title{flow mat}
\usage{
as.flowmat(x, grp_col, jobname_col, cmd_col, ...)

is.flowmat(x)

to_flowmat(x, ...)

\method{to_flowmat}{list}(x, samplename, ...)

\method{to_flowmat}{data.frame}(x, ...)

\method{to_flowmat}{flow}(x, ...)
}
\arguments{
\item{x}{a data.frame or path to file with flow details in it.}

\item{grp_col}{column used for grouping, default samplename.}

\item{jobname_col}{column specifying jobname, default jobname}

\item{cmd_col}{column specifying commands to run, default cmd}

\item{...}{not used}

\item{samplename}{character of length 1 or that of nrow(x)}

\item{x}{a named list OR vector. Where name corresponds to the jobname and value is a vector of commands to run}

\item{...}{not used}
}
\description{
as.flowmat(): reads a file and checks for required columns. If x is data.frame checks for required columns.

Taking in a named list and returns a two columns data.frame
}

