% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_module.R
\name{TransformUI}
\alias{TransformUI}
\alias{Transform}
\title{Edit flow parameter transformation and description}
\usage{
TransformUI(id)

Transform(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Edit flow parameter transformation and description

Transform module server function
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)

if (interactive()){
  
  ui <- dashboardPage(
    dashboardHeader(title = "Transform"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      TransformUI("module")
    )
  )
  
  server <- function(input, output, session) {
    rval <- reactiveValues()
    observe({
      utils::data("GvHD", package = "flowCore")
      rval$gating_set <- GatingSet(GvHD)
    })
    res <- callModule(Transform, "module", rval = rval)
  }
  
  shinyApp(ui, server)
  
}
}
}
