% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{update.flow}
\alias{update.flow}
\title{update.flow}
\usage{
\method{update}{flow}(object, child, ...)
}
\arguments{
\item{object}{flowr object}

\item{child}{child}

\item{...}{not used}
}
\value{
flowr object
}
\description{
update.flow
}
