% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\name{update_flow_det}
\alias{update_flow_det}
\title{update_flow_det}
\usage{
update_flow_det(wd, mat_cmd)
}
\arguments{
\item{wd}{flow working directory}

\item{mat_cmd}{a table with details about cmd files}
}
\description{
update_flow_det
}
\details{
Get the flow_det files from wd, and update it with new statuses.
}

