% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submit-job.R
\name{whisker_render}
\alias{whisker_render}
\title{a wrapper around whisker.render with some sugar on it...}
\usage{
whisker_render(template, data)
}
\description{
a wrapper around whisker.render with some sugar on it...
}

