% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write-sheet.R
\name{write_sheet}
\alias{write_sheet}
\title{A simple wrapper around write.table with some sugar.}
\usage{
write_sheet(x, file)
}
\description{
A simple wrapper around write.table with some sugar.
}

