% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zenith_module.R
\name{ZenithUI}
\alias{ZenithUI}
\alias{Zenith}
\title{Analyse cell metabolism using the ZeNITH method}
\usage{
ZenithUI(id)

Zenith(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Analyse cell metabolism using the ZeNITH method

Zenith module server function
}
