% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flowdef.R
\name{as.flowdef}
\alias{as.flowdef}
\alias{is.flowdef}
\title{flow definition}
\usage{
as.flowdef(x, ...)

is.flowdef(x)
}
\arguments{
\item{x}{can be a data.frame or a path for a flow definition file}

\item{...}{passed onto check.flowdef}
}
\description{
Reeading a flow definition file and checking it.
}

