% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.r
\name{bayesass}
\alias{bayesass}
\title{Run BayesAss.}
\usage{
bayesass(iter = 1e+07, exe = "E:/flow/BA3Windows64/BA3.exe",
  infile = "infile_bayesass.txt", wd = getwd())
}
\arguments{
\item{iter}{Number of interations}

\item{exe}{Path to BayesAss executable file}

\item{infile}{Path to input gene data file}

\item{wd}{Path to working directory for file export}
}
\description{
Run BayesAss.
}

