% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.r
\name{bayesass}
\alias{bayesass}
\title{Run BayesAss.}
\usage{
bayesass(infile = "infile_bayesass.txt", outfile = "outfile_bayesass.txt",
  iter = 1e+07, burn = 1e+06, interval = 1000, seed = 10,
  other = NULL, exe = "E:/flow/BA3Windows64/BA3.exe")
}
\arguments{
\item{infile}{Path to input gene data file}

\item{outfile}{Output filename including .txt}

\item{iter}{Number of iterations (integer)}

\item{burn}{Burn-in iterations (integer)}

\item{interval}{Sampling interval (integer)}

\item{seed}{Random starting seed (integer)}

\item{other}{Any other args to BA3.exe}

\item{exe}{Path to BayesAss executable file}
}
\description{
Run BayesAss.
}

