% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\alias{check.flowdef}
\alias{check.flowmat}
\title{Check consistency of flowdef and flowmat}
\usage{
check(x, ...)

\method{check}{flowmat}(x, ...)

\method{check}{flowdef}(x, ...)
}
\arguments{
\item{x}{a flowdef or flowmat object}

\item{...}{suppled to \code{check.classname} function}
}
\description{
check consistency of objects
Currently checks objects S3 flowdef, flowmat
}

