% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_args}
\alias{check_args}
\title{Assert none of the arguemnts of a function are null.}
\usage{
check_args(ignore, select)
}
\arguments{
\item{ignore}{optionally ignore a few variables for checking.}

\item{select}{optionally only check a few variables of the function.}
}
\description{
Checks all the arguments in the parent function and makes sure that none of them
are NULL
}
\details{
This function has now been moved to params package.
}

