% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_module.R
\name{CleanUI}
\alias{CleanUI}
\title{Identify and remove anomalies}
\usage{
CleanUI(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
Identify and remove anomalies
}
\examples{
\dontrun{
if (interactive()){

  ui <- dashboardPage(
    dashboardHeader(title = "flowAI"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      CleanUI("module")
    )
  )

  server <- function(input, output, session) {
    rval <- reactiveValues()
    observe({
      utils::data("Bcells", package = "flowAI")
      rval$gating_set <- flowWorkspace::GatingSet(Bcells)
    })
    res <- callModule(Clean, "module", rval = rval)
  }
  
  shinyApp(ui, server)

}}
}
