% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_module.R
\name{ClusteringUI}
\alias{ClusteringUI}
\alias{Clustering}
\title{Identify clusters}
\usage{
ClusteringUI(id)

Clustering(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Identify clusters

Clustering module server function
}
