% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{cmds_to_flow}
\alias{cmds_to_flow}
\title{convert command matrix to specific flow}
\usage{
cmds_to_flow(cmd.list, samplename = "", infomat, q_obj = queue(type = "lsf",
  queue = "normal"), flowname = "stage2", execute = FALSE,
  flow_base_path = "~/tmp")
}
\description{
convert commands to flow
}

