% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flow.R
\name{cmds_to_flow}
\alias{cmds_to_flow}
\title{to_flow}
\usage{
cmds_to_flow(cmd.list, samplename = "", infomat, q_obj = queue(type = "lsf",
  verbose = FALSE), flowname = "stage2", execute = FALSE,
  flow_run_path = "/scratch/iacs/flow_pipe/tmp")
}
\arguments{
\item{x}{\link{list} of commands}
}
\description{
Create a \link{flow} object from a list of commands
}

