% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation_module.R
\name{compensationUI}
\alias{compensationUI}
\title{compensationUI and compensation}
\usage{
compensationUI(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module that deals with compensation
}
