% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zz-create-jobs-mat.R
\name{create_jobs_mat}
\alias{create_jobs_mat}
\title{create_jobs_mat}
\usage{
create_jobs_mat(x)
}
\arguments{
\item{x}{a \link{flow} object.}
}
\description{
create_jobs_mat
}
\details{
This create a table similar to flowdef, except it has
jobids and previous jobids instead of only names.
This helps in creating the diagram for the flow.
Currently the only way to create a diagram is from a flow object,
by calling the function plot_flow.

In future plot_flow may be able to accept a flowdef only.
}
\examples{
\dontrun{
create_jobs_mat(x = x)}
}
\keyword{internal}

