% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/class-funcs.R
\name{create_queue_cmd}
\alias{create_queue_cmd}
\title{create_queue_cmd}
\usage{
create_queue_cmd(j_obj, file, index, ...)
}
\arguments{
\item{j_obj}{}

\item{file}{This is the path to the file to run}

\item{index}{}

\item{...}{}
}
\description{
.create_queue_cmd
}
\examples{
.create_queue_cmd(j_obj = j_obj, file = file, index = index, ... = ...)
}

