% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-funcs.R
\name{create_queue_cmd}
\alias{create_queue_cmd}
\title{create_queue_cmd}
\usage{
create_queue_cmd(j_obj, file, index, ...)
}
\arguments{
\item{j_obj}{object of class \link{job}}

\item{file}{This is the path to the file to run}

\item{index}{among cmds defined in \code{j_obj}, which index does this \code{file} belong to. A numeric vector of length 1. This is to fetch dependency from previous job.}

\item{...}{Not used}
}
\description{
This is a flow interal functions used to create a command used to submit jobs to the cluster.
}
\examples{
\dontrun{
.create_queue_cmd(j_obj = j_obj, file = file, index = index, ... = ...)
}
}
\keyword{internal}

