% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed_module.R
\name{dimRed}
\alias{dimRed}
\title{dim_reduction server function}
\usage{
dimRed(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
dim_reduction server function
}
