% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduction_module.R
\name{dimRed}
\alias{dimRed}
\title{dim_reduction server function}
\usage{
dimRed(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
dim_reduction server function
}
