% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proc.errors.R
\name{error}
\alias{error}
\title{Error Handler}
\usage{
error(x)
}
\arguments{
\item{x}{this is a string(s) with short error summary}
}
\description{
function to handle all error descriptions
}
\keyword{internal}

