% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\name{fetch}
\alias{fetch}
\alias{fetch_conf}
\alias{fetch_pipes}
\title{A generic functions to search for files}
\usage{
fetch(x, pattern, places, urls, verbose = FALSE)

fetch_pipes()

fetch_conf(x = "flowr.conf", places, ...)
}
\arguments{
\item{x}{name of the file to search for}

\item{places}{places (paths) to look for it. Its best to use the defaults}

\item{urls}{urls to look for, works well for pipelines.}

\item{verbose}{be chatty?}
}
\description{
A generic functions to search for files
}
\details{
sequence: the later overrides former if parameter name matches
~/                                 ## home dir
~/flowr/conf                       ## flowr default home
MYRLIB/flowr/conf folder           ## flow pipeline folder

Looks at: github repo: ngsflows/pipelines
}
\examples{
{
search_conf("torque.sh")
}
}

