% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch.r
\name{fetch}
\alias{fetch}
\alias{fetch_conf}
\alias{fetch_pipes}
\title{A generic functions to search for files}
\usage{
fetch(x, places, urls, verbose = FALSE)

fetch_pipes(x, places, last_only = FALSE,
  urls = get_opts("flowr_pipe_urls"), silent = FALSE, ask = TRUE)

fetch_conf(x = "flowr.conf", places, ...)
}
\arguments{
\item{x}{name of the file to search for}

\item{places}{places (paths) to look for it. Its best to use the defaults}

\item{urls}{urls to look for, works well for pipelines.}

\item{verbose}{be chatty?}

\item{silent}{applicable to fetch_pipes() only. Throw if no such pipe is available}

\item{ask}{ask before downloading or copying, not used !}

\item{...}{not used}
}
\description{
These functions help in searching for specific files in the user space.

MYRLIB/flowr/conf folder           ## flowr/ngsflows internal default configurations
~/flowr/conf                       ## flowr default home

fetch_pipes(): Looks at: github repo: ngsflows/pipelines

fetch_conf(): Searching for .conf files in various places
}
\examples{
{
fetch_conf("torque.sh")
}
}

