% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{find_all_parent_gates}
\alias{find_all_parent_gates}
\title{find, in a recursive manner, all parent gates of a given Gate xml_node
from a FlowJO wsp}
\usage{
find_all_parent_gates(x)
}
\arguments{
\item{x}{a Gate xml_node from a FlowJO wsp}
}
\description{
find, in a recursive manner, all parent gates of a given Gate xml_node
from a FlowJO wsp
}
