% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR_server.R
\name{flowR_server}
\alias{flowR_server}
\title{flowR server function}
\usage{
flowR_server(session, input, output, user_module_name = NULL)
}
\arguments{
\item{session}{shiny session}

\item{input}{shiny input}

\item{output}{shiny output}

\item{user_module_name}{name of a user shiny module (server function) to integrate to the app}
}
\description{
flowR server function
}
