% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flowsets_module.R
\name{FlowsetsUI}
\alias{FlowsetsUI}
\alias{Flowsets}
\title{Manage and visualize flow-sets and their hierarchy}
\usage{
FlowsetsUI(id)

Flowsets(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactive values object \code{rval}
}
\description{
Manage and visualize flow-sets and their hierarchy

Flowsets module server function
}
