% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsetsUI_module.R
\name{flowsetsUI}
\alias{flowsetsUI}
\alias{flowsets}
\title{flowsetsUI and flowsets}
\usage{
flowsetsUI(id)

flowsets(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
A shiny Module that imports a flow set and gates from fcs files and a workspace
}
