% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating_module.R
\name{GatingUI}
\alias{GatingUI}
\alias{Gating}
\title{Edit, visualize and show statistics from a gating hierarchy}
\usage{
GatingUI(id)

Gating(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Edit, visualize and show statistics from a gating hierarchy

Gating module server function
}
