% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating2_module.R, R/gating_module.R
\name{gating2}
\alias{gating2}
\alias{gatingUI}
\alias{gating}
\title{gating server function}
\usage{
gating2(input, output, session, rval)

gatingUI(id)

gating(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}

\item{id}{shiny id}
}
\value{
The updated reactiveValues object \code{rval}

The updated reactiveValues object \code{rval}
}
\description{
A shiny Module that deals with gating
}
