% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowmat-funcs.R
\name{generate_flow_code}
\alias{generate_flow_code}
\title{generate_flow_code}
\usage{
generate_flow_code(x, ...)
}
\arguments{
\item{x}{flow object}

\item{...}{currently ignored}
}
\description{
generate_flow_code
}
\examples{
\dontrun{
generate_flow_code(x = x)
}
}
\keyword{internal}

