% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/status-funcs.R
\name{get_flow_status}
\alias{get_flow_status}
\alias{status}
\title{get_flow_status}
\usage{
get_flow_status(x, cores = 6, out_format = "markdown",
  get_mem_usage = TRUE)
}
\arguments{
\item{x}{path to the flow; may be without the uuid}

\item{cores}{number of cores to use}

\item{out_format}{passed onto knitr:::kable. supports: markdown, rst, html...}

\item{get_mem_usage}{under progress, whether to extract mem_usage of jobs}
}
\description{
get_flow_status
}
\examples{
\dontrun{
get_flow_status(x = x, cores = 6)
## an example for running from terminal
flowr status x=path_to_flow_directory cores=6
}
}

