% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-parse-resources.R
\name{get_resources}
\alias{get_resources}
\title{Extract resources used by each job of a flow
get_resources currenty this only works on LSF}
\usage{
get_resources(x, odir, ...)
}
\arguments{
\item{x}{A character vector of lenth 1. This may be a parent level folder with directories with multiple flow runs.}

\item{odir}{Output directory to save the results}

\item{\dots}{other arguments sent to \link{get_resources_lsf}}
}
\description{
Extract resources used by each job of a flow
get_resources currenty this only works on LSF
}
\details{
If \code{x} is a parent level folder, 
then resources are summarized for all its child folders.
}
\examples{
\dontrun{
get_resources(x = x, odir = ~/tmp)
}
}
