% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\name{get_status}
\alias{get_status}
\alias{get_status.character}
\alias{get_status.data.frame}
\alias{get_status.flow}
\title{get_status}
\usage{
get_status(x, ...)

\method{get_status}{character}(x, out_format = "markdown", ...)

\method{get_status}{data.frame}(x, ...)

\method{get_status}{flow}(x, out_format = "markdown", ...)
}
\arguments{
\item{x}{flow, flow_det or a path to flow run.}

\item{...}{not used}

\item{out_format}{format in which status is to be printed. Default is markdown}
}
\description{
get_status

Uses the trigger column to get the exit status and updates the data.frame.
}

