% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R
\name{get_unique_id}
\alias{get_unique_id}
\title{get_unique_id}
\usage{
get_unique_id(prefix = "id", suffix = "", random_length = 8)
}
\arguments{
\item{prefix}{Default \code{id}. Character string to be added in the front.}

\item{suffix}{Default ''. Character string to be added in the end.}

\item{random_length}{Integer, defaults to 8. In our opinion 8 serves well, providing 'uniqueness' and not being much of a eyesore.}
}
\description{
get_unique_id
}
\examples{
\dontrun{
get_unique_id(base = id, random_length = 8)}
}
\keyword{internal}

