% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xtab.R
\name{get_xtab}
\alias{get_xtab}
\title{After binning the data, make a table with the nearly empty bins combined into a 'leftover bin'. often dramatically reduces size of the table without much information loss.}
\usage{
get_xtab(data_binned, w, C)
}
\arguments{
\item{data_binned}{output of make_bins}

\item{w}{the 'vacancy rate', a number between 0 and 1. The final table will have no column (bin) with a greater proportion than w of donors with zero events in that bin.}

\item{C}{The row names - donor sample type}
}
\value{
The reduced data matrix. Each row (donor) sums to one. The first column is the number of 'leftover' events, i.e. the events that don't fit into the bins that are common to most samples.
}
\description{
After binning the data, make a table with the nearly empty bins combined into a 'leftover bin'. often dramatically reduces size of the table without much information loss.
}

