% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportSpill_module.R
\name{ImportSpillUI}
\alias{ImportSpillUI}
\alias{ImportSpill}
\title{Import spillover matrix}
\usage{
ImportSpillUI(id)

ImportSpill(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The Importd matrix
}
\description{
Import spillover matrix

ImportSpill server function
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)

if (interactive()){
 
  ui <- dashboardPage(
    dashboardHeader(title = "ImportSpill"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      box(title="Import", width = NULL, height = NULL,
          ImportSpillUI("module")
          )
    )
  )

  server <- function(input, output, session) {
    rval <- reactiveValues()
    observe({
      utils::data("GvHD", package = "flowCore")
      rval$gating_set <- GatingSet(GvHD)
    })

    res <- callModule(ImportSpill, "module", rval = rval)
  }

  shinyApp(ui, server)

}
}
}
