% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_module.R, R/metadata_module.R
\name{importUI}
\alias{importUI}
\alias{import}
\alias{metadata}
\title{importUI and import}
\usage{
importUI(id)

import(input, output, session)

metadata(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object

The updated reactiveValues object \code{rval}
}
\description{
A shiny Module that imports data and builds flow-sets
}
