% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importUI_module.R, R/metadataUI_module.R
\name{importUI}
\alias{importUI}
\alias{import}
\alias{metadata}
\title{importUI and import}
\usage{
importUI(id)

import(input, output, session)

metadata(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"

a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
A shiny Module that imports a flow set and gates from fcs files and a workspace
}
