% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_module.R
\name{ImportUI}
\alias{ImportUI}
\alias{Import}
\title{Import data and build a GatingSet}
\usage{
ImportUI(id)

Import(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{a reactivevalues object (can be empty)}
}
\value{
The input reactivevalues object 'rval' with updated elements :
\describe{
  \item{gating_set_list}{list of GatingSet objects loaded}
}
}
\description{
Import data and build a GatingSet

Import module server function
}
