% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/class-def.R
\name{job}
\alias{job}
\title{job class}
\usage{
job(cmds = "", base_path = "", parent_flow = "", name = "myjob",
  q_obj = new("queue"), previous_job = "", cpu = 1, memory, walltime,
  submission_type = c("scatter", "serial"), status = "", script = "",
  dependency_type = c("none", "gather", "serial", "burst"), ...)
}
\arguments{
\item{cmds}{the commands to run}

\item{base_path}{base path}

\item{parent_flow}{parent flow}

\item{name}{name of the job}

\item{q_obj}{queue object}

\item{previous_job}{character vector of previous job. If this is the first job, one can leave this empty, NA, NULL or ''. In future this could specify multiple previous jobs.}

\item{cpu}{no of cpu's reserved}

\item{submission_type}{submission type: A character with values: scatter, serial. Scatter means all the 'cmds' would be run in parallel as seperate jobs. Serial, they would combined into a single job and run one-by-one.}

\item{status}{status [ignore]. this is used internally to update status of the job.}

\item{dependency_type}{depedency type. One of none, gather, serial, burst. If previous_job is specified, then this would not be 'none'. [Required]}

\item{...}{other passed onto object creation. Example: memory, walltime, cpu}
}
\description{
job class
}
\examples{
q_obj <- queue(type="torque")
j_obj <- job(q_obj=q_obj, cmd = "sleep 2", cpu=1)
}

