% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kill-flow.R
\name{kill}
\alias{kill}
\alias{kill.character}
\alias{kill.flow}
\title{kill}
\usage{
kill(x, ...)

\method{kill}{character}(x, force = FALSE, ...)

\method{kill}{flow}(x, kill_cmd, jobid_col = "job_sub_id", ...)
}
\arguments{
\item{x}{either path to flow [character] or fobj object of class \link{flow}}

\item{...}{not used}

\item{force}{When killing multiple flows, force is neccesary. This makes sure multiple flows are killed by accident.}

\item{kill_cmd}{The command used to kill. Default is 'bkill' (LSF). One can used qdel for 'torque', 'sge' etc.}

\item{jobid_col}{Advanced use. The column name in 'flow_details.txt' file used to fetch jobids to kill}
}
\description{
kill

works on flow_path. Reads flow object and calls kill.flow()

works on flow object
}
\examples{
\dontrun{
## example for terminal
## flowr kill_flow x=path_to_flow_directory
}
}

