% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/status-funcs.R
\name{kill_flow}
\alias{kill_flow}
\title{kill_flow}
\usage{
kill_flow(x, wd, fobj, kill_cmd = "bkill", jobid_col = "job_sub_id")
}
\arguments{
\item{x}{either path to flow [character] or fobj object of class \link{flow}}

\item{wd}{path to a specific which needs to be killed}

\item{fobj}{a object of class \link{flow}}

\item{kill_cmd}{The command used to kill. Default is 'bkill' (LSF). One can used qdel for 'torque', 'sge' etc.}

\item{jobid_col}{Advanced use. The column name in 'flow_details.txt' file used to fetch jobids to kill}
}
\description{
kill_flow
}
\examples{
\dontrun{
## example for terminal
flowr kill_flow wd=path_to_flow_directory
}
}

