% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\name{load_conf}
\alias{load_conf}
\title{load a configuration file into the environment}
\usage{
load_conf(x, chk = TRUE, ...)
}
\arguments{
\item{x}{path to a configuration file}

\item{chk}{check the file after loading?}

\item{...}{Not used}
}
\description{
load a configuration file into the environment
}

