% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf.R
\name{load_conf}
\alias{load_conf}
\title{Fetching and loading configuration files.}
\usage{
load_conf(x, chk = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{path to a configuration file}

\item{chk}{check the file after loading?}

\item{...}{Not used}
}
\description{
These are two column tab seperated files (namely name and value).
}
\seealso{
\link{get_opts}
}

