% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata_module.R
\name{MetadataUI}
\alias{MetadataUI}
\alias{Metadata}
\title{Import, export and edit metadata associated with a GatingSet}
\usage{
MetadataUI(id)

Metadata(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Import, export and edit metadata associated with a GatingSet

metadata module server function
}
