% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modules_module.R
\name{ModulesUI}
\alias{ModulesUI}
\alias{Modules}
\title{Search, select and load modules}
\usage{
ModulesUI(id)

Modules(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{reactivevalues object with the following elements :
\describe{
  \item{gating_set}{: a GatingSet object}
  \item{menu_elements}{: Named list. Names should correspond to the names of available modules.}
}}
}
\value{
The input reactivevalues object 'rval' with updated elements :
\describe{
  \item{modules}{: names of selected modules}
}
}
\description{
Search, select and load modules

server function for the shiny module Modules
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)
library(flowR)
library(DT)

if (interactive()){
  
  ui <- dashboardPage(
    dashboardHeader(title = "Modules"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      ModulesUI("module")
    )
  )
  
  server <- function(input, output, session) {
    rval <- reactiveValues(menu_elements = list("Modules" = NULL))
    rval <- callModule(Modules, "module", rval = rval)
  }
  
  shinyApp(ui, server)
  
}
}
}
