% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization_module.R
\name{NormalizationUI}
\alias{NormalizationUI}
\title{Normalize data from beads}
\usage{
NormalizationUI(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
Normalize data from beads
}
\examples{
\dontrun{
if (interactive()){

  ui <- dashboardPage(
    dashboardHeader(title = "premessa"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      NormalizationUI("module")
    )
  )

  server <- function(input, output, session) {
    rval <- reactiveValues()
    observe({
     #files <- path of you're files 
     fs <- read.ncdfFlowSet(files = files)
      rval$gating_set <- flowWorkspace::GatingSet(fs)
    })
    res <- callModule(Normalization, "module", rval = rval)
  }
  
  shinyApp(ui, server)

}}
}
