% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseGroupNodes}
\alias{parseGroupNodes}
\title{Return the name and IDs of samples in a  GroupNode xml_node
from a flowJO wsp}
\usage{
parseGroupNodes(x)
}
\arguments{
\item{x}{a xml_document object}
}
\description{
Return the name and IDs of samples in a  GroupNode xml_node
from a flowJO wsp
}
