% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{parseSampleNodes}
\alias{parseSampleNodes}
\title{Return the name and ID of a SampleNode xml_node 
from a FlowJO wsp}
\usage{
parseSampleNodes(x)
}
\arguments{
\item{x}{a xml_document object}
}
\description{
Return the name and ID of a SampleNode xml_node 
from a FlowJO wsp
}
