% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_flow}
\alias{.plot_flow_dat_type1}
\alias{plot_flow}
\alias{plot_flow.flow}
\alias{plot_flow.list}
\title{plot_flow}
\usage{
plot_flow(x, ...)

\method{plot_flow}{flow}(x, detailed = TRUE, type = c("1", "2"),
  pdf = FALSE, pdffile = sprintf("\%s.pdf", x@name), ...)

\method{plot_flow}{list}(x, ...)

.plot_flow_dat_type1(x, detailed = FALSE, pdf = FALSE,
  pdffile = sprintf("flow_details.pdf"), width, height, ...)
}
\arguments{
\item{x}{Object of class \code{flow}}

\item{...}{experimental}

\item{detailed}{include some details}

\item{type}{1 is original, and 2 is a elipse with less details}

\item{pdf}{create a pdf instead of plotting interactively}

\item{pdffile}{output file name for the pdf file}

\item{x}{a data.frame as given out create_jobs_mat()}
}
\description{
plot the flow object
}
\examples{
qobj = queue(type="lsf")
cmds = rep("sleep 5", 10)
jobj1 <- job(q_obj=qobj, cmd = cmds, submission_type = "scatter", name = "job1")
jobj2 <- job(q_obj=qobj, name = "job2", cmd = cmds, submission_type = "scatter",
             dependency_type = "serial", previous_job = "job1")
fobj <- flow(jobs = list(jobj1, jobj2))
plot_flow(fobj)

### Gather: many to one relationship
jobj1 <- job(q_obj=qobj, cmd = cmds, submission_type = "scatter", name = "job1")
jobj2 <- job(q_obj=qobj, name = "job2", cmd = cmds, submission_type = "scatter",
             dependency_type = "gather", previous_job = "job1")
fobj <- flow(jobs = list(jobj1, jobj2))
plot_flow(fobj)

### Burst: one to many relationship
jobj1 <- job(q_obj=qobj, cmd = cmds, submission_type = "serial", name = "job1")
jobj2 <- job(q_obj=qobj, name = "job2", cmd = cmds, submission_type = "scatter",
             dependency_type = "burst", previous_job = "job1")
fobj <- flow(jobs = list(jobj1, jobj2))
plot_flow(fobj)
}

