% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_flow}
\alias{plot_flow}
\alias{plot_flow.list}
\alias{plot_flow.flow}
\alias{plot}
\alias{plot_flow.character}
\alias{plot_flow.flowdef}
\alias{plot.flowdef}
\alias{plot.flow}
\title{Plot a clean and scalable flowchart describing the (work)flow}
\usage{
plot_flow(x, ...)

\method{plot_flow}{flow}(x, ...)

\method{plot_flow}{list}(x, ...)

\method{plot_flow}{character}(x, ...)

\method{plot_flow}{flowdef}(x, detailed = TRUE, type = c("1", "2"), pdf = FALSE, pdffile, ...)

\method{plot}{flowdef}(x, detailed = TRUE, type = c("1", "2"), pdf = FALSE, pdffile, ...)

\method{plot}{flow}(x, ...)
}
\arguments{
\item{x}{Object of class \code{flow}, or a list of flow objects or a flowdef}

\item{...}{experimental and only for advanced use.}

\item{detailed}{include submission and dependency types in the plot [TRUE]}

\item{type}{1 is original, and 2 is a ellipse with less details [1]}

\item{pdf}{create a pdf instead of plotting interactively [FALSE]}

\item{pdffile}{output file name for the pdf file. [\code{flow_path/flow_details.pdf}]}
}
\description{
Plot a flowchart using a flow object or flowdef
}
\examples{
qobj = queue(type="lsf")
cmds = rep("sleep 5", 10)
jobj1 <- job(q_obj=qobj, cmd = cmds, submission_type = "scatter", name = "job1")
jobj2 <- job(q_obj=qobj, name = "job2", cmd = cmds, submission_type = "scatter",
             dependency_type = "serial", previous_job = "job1")
fobj <- flow(jobs = list(jobj1, jobj2))
plot_flow(fobj)

### Gather: many to one relationship
jobj1 <- job(q_obj=qobj, cmd = cmds, submission_type = "scatter", name = "job1")
jobj2 <- job(q_obj=qobj, name = "job2", cmd = cmds, submission_type = "scatter",
             dependency_type = "gather", previous_job = "job1")
fobj <- flow(jobs = list(jobj1, jobj2))
plot_flow(fobj)

### Burst: one to many relationship
jobj1 <- job(q_obj=qobj, cmd = cmds, submission_type = "serial", name = "job1")
jobj2 <- job(q_obj=qobj, name = "job2", cmd = cmds, submission_type = "scatter",
             dependency_type = "burst", previous_job = "job1")
fobj <- flow(jobs = list(jobj1, jobj2))
plot_flow(fobj)

}
