% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_gs}
\alias{plot_gs}
\title{Plot a GatingSet}
\usage{
plot_gs(
  gs,
  df = NULL,
  sample = NULL,
  subset = NULL,
  spill = NULL,
  metadata = NULL,
  gate_name = NULL,
  Ncells = NULL,
  plot_type = "hexagonal",
  plot_args = list(),
  options = list()
)
}
\arguments{
\item{gs}{a GatingSet}

\item{df}{a data.frame with plot data resulting from a call of \code{get_plot_data}. 
Supersedes parameters 'gs', 'sample', 'spill', 'metadata'}

\item{sample}{Names of samples from the GatingSet 
(as returned by \code{pData(gs)$name})}

\item{subset}{Names of subsets from the GatingSet 
(as returned by \code{gs_get_pop_paths(gs)})}

\item{spill}{spillover matrix. If NULL, uncompensated data is used for gating and plotting.}

\item{metadata}{a data.frame containing metadata associated to samples.
Must have a column \code{name} used for mapping.}

\item{gate_name}{Names of the gates to add to the plot (if it is compatible with plot parameters).
Ignored if NULL.}

\item{plot_type}{name of the plot type}

\item{plot_args}{list of plot parameters passed to the plot function. 
Plot parameters depend on the plot type selected.}

\item{options}{list of plot format options passed to \code{format_plot()}}
}
\value{
a plot
}
\description{
Plot a GatingSet
}
