% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_stat}
\alias{plot_stat}
\title{Plot summary statistics from a gating set}
\usage{
plot_stat(df = NULL, gs, sample, subset, yvar, type = "bar",
  metadata = NULL, color_var = "subset", axis_labels = NULL,
  transformation = NULL, spill = NULL,
  default_trans = identity_trans(), scale_values = FALSE,
  free_y_scale = TRUE, max_scale = 0, facet_vars = NULL,
  group_var = "subset", expand_factor = 0.2, stat_function = "mean",
  show.legend = TRUE, y_trans = NULL, strip.text.y.angle = 0,
  theme_name = "theme_gray", Rowv = TRUE, Colv = TRUE)
}
\arguments{
\item{df}{data.frame with columns \code{name} and \code{subset} 
containing sample and subset names respectively and
columns with plot variables. Ignored if \code{NULL}}

\item{gs}{a gating set}

\item{sample}{sample names}

\item{subset}{subset names}

\item{yvar}{y variable}

\item{type}{plot type : either "bar", "tile" or "heatmap"}

\item{metadata}{data.frame with metadata}

\item{color_var}{color variable}

\item{axis_labels}{Named vector with axis labels. 
Vector names should correspond to variable names}

\item{transformation}{Named list of \code{trans} objects. 
List names should correspond to variable names.}

\item{spill}{spillover matrix. If NULL, uncompensated data is used both for gating and plotting.}

\item{scale_values}{logical. Scale values for each y variable}

\item{free_y_scale}{logical. Free y scale (use a different scale for each y variable)}

\item{max_scale}{maximum scale for absolute y values. Used only if \code{scale_values = TRUE}}

\item{facet_vars}{variables used to facet plots along the x axis}

\item{group_var}{group variable}

\item{expand_factor}{numeric value expanding y scale. Used only if \code{free_y_scale = FALSE}}

\item{stat_function}{Name of the summary statistics function used}

\item{show.legend}{logical. Should plot legend be displayed?}

\item{y_trans}{default \code{trans} object to be used if \code{transformation} is NULL.}

\item{strip.text.y.angle}{angle of y strip labels}

\item{theme_name}{ggplot2 theme name ('theme_gray' by default)}

\item{Rowv}{Cluster rows using hierarchical clustering?}

\item{Colv}{Cluster columns using hierarchical clustering?}

\item{type}{plot type}

\item{metadata}{data.frame with metadata.}
}
\value{
a ggplot object
}
\description{
Plot summary statistics from a gating set
}
