% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowR.R
\name{plot_stat}
\alias{plot_stat}
\title{#' Plot data and gates from a gating set
#' @description  Plot data and gates from a gating set
#' @param df data.frame with columns \code{name} and \code{subset} 
#' containing sample and subset names respectively and
#' columns with plot variables. Ignored if \code{NULL}
#' @param gs a gating set
#' @param sample sample names
#' @param subset subset names
#' @param xvar x variable
#' @param yvar y variable
#' @param axis_labels Named vector with axis labels. 
#' Vector names should correspond to variable names
#' @param color_var color variable
#' @param data_range Named list of data ranges. 
#' List names should correspond to variable names.
#' @param min_value numeric. Minimal value of x and y axis. Ignored if \code{NULL}.
#' @param gate A character vector with subset names (i.e elements of \code{getNodes(gs)}) or a list of gates 
#' (as returned by \code{get_gates_from_gs()}). 
#' Gates from classes \code{polygonGate}, \code{rectangleGate} and \code{ellipsoidGate} are supported.
#' If \code{xvar} and \code{yvar} are \code{NULL}, the first gate will be used to set plot variables.
#' @param polygon_gate a matrix with columns \code{x} and \code{y} containing path coordinates.
#' @param type plot type : either "hexagonal", "histogram", "dots" or "contour"
#' @param bins bins parameter
#' @param alpha alpha parameter
#' @param size size parameter
#' @param transformation  Named list of \code{trans} objects. 
#' List names should correspond to variable names.
#' @param default_trans default \code{trans} object to be used if \code{transformation} is NULL.
#' @param spill spillover matrix. If NULL, uncompensated data is used both for gating and plotting.
#' @param metadata data.frame with metadata.
#' @param facet_vars variables used to facet plots along the x axis
#' @param yridges_var y variables used in \code{geom_density_ridges()}. Used only if \code{ridges} is TRUE.
#' @param norm_density normalize maximum density to 1
#' @param smooth use \code{geom_density()} instead of \code{geom_histogram()}. Only used if \code{type = "histogram"}.
#' @param ridges use \code{geom_density_ridges()}. Only used if \code{type = "histogram"} and \code{smooth = TRUE}.
#' @param bw deprecated
#' @param show.legend logical. Should plot legend be displayed?
#' @param legend.position legend position. "right" by default
#' @param use_log10_count logical. Use log-10 transform for color scale if \code{type = "hexagonal"}
#' @param theme_name ggplot2 theme name ('theme_gray' by default)
#' @return a ggplot object
#' @import flowWorkspace
#' @import ggplot2
#' @import ggridges
#' @import viridis
#' @import scales
plot_gs <- function(df = NULL,
                    gs = NULL, 
                    sample,
                    subset,
                    xvar = NULL,
                    yvar = NULL,
                    axis_labels = NULL,
                    color_var = NULL, 
                    color_var_name = NULL,
                    data_range = NULL,
                    min_value = NULL,
                    gate=NULL, 
                    polygon_gate = NULL,
                    type = "hexagonal", 
                    bins = 100,
                    alpha = 0.1,
                    size = 0.1,
                    transformation = NULL,
                    default_trans = identity_trans(),
                    spill = NULL,
                    metadata = NULL,
                    facet_vars = NULL,
                    #group_var = "name",
                    yridges_var = "name",
                    norm_density = TRUE,
                    smooth = FALSE,
                    ridges = FALSE,
                    bw = 0.1,
                    show.legend = TRUE,
                    legend.position = "right",
                    use_log10_count = TRUE,
                    theme_name = "theme_gray"
                    ){
  
  
  if(!is.null(color_var)){
    if(color_var %in% c("", "none")){
      color_var <- NULL
    }
  }
  
  if(!is.numeric(alpha)){
    alpha = 0.1
  }
  
  if(!is.numeric(size)){
    size = 0.1
  }
  
  if(!is.numeric(bins)){
    bins = 100
  }
  
  
  theme_function <- function(...){
    do.call(theme_name, list(...))
  }
  
  if(!is.null(gs)){
    idx <- match(sample, pData(gs)$name)
  }
  
  if(!is.null(gate)){
    if(class(gate) == "character"){
      if(!is.null(gs)){
        
        gate_int <- lapply(gate, function(x){
          if(x!="root"){
            getGate(gs[[idx[1]]], x)
          }else{
            NULL
          }})
        names(gate_int) <- gate
        gate <- gate_int
      }else{
        gate <- NULL
      }
      
    }else if(grep("Gate", class(gate)) >0){
      gate <- list(gate)
    }
  }}
\usage{
plot_stat(df = NULL, gs, sample, subset, yvar, type = "bar",
  metadata = NULL, color_var = "subset", axis_labels = NULL,
  transformation = NULL, spill = NULL,
  default_trans = identity_trans(), scale_values = FALSE,
  free_y_scale = TRUE, max_scale = 0, facet_vars = NULL,
  group_var = "subset", expand_factor = 0.2, stat_function = "mean",
  show.legend = TRUE, y_trans = NULL, strip.text.y.angle = 0,
  theme_name = "theme_gray", Rowv = TRUE, Colv = TRUE)
}
\arguments{
\item{df}{data.frame with columns \code{name} and \code{subset} 
containing sample and subset names respectively and
columns with plot variables. Ignored if \code{NULL}}

\item{gs}{a gating set}

\item{sample}{sample names}

\item{subset}{subset names}

\item{yvar}{y variable}

\item{type}{plot type : either "bar", "tile" or "heatmap"}

\item{metadata}{data.frame with metadata}

\item{color_var}{color variable}

\item{axis_labels}{Named vector with axis labels. 
Vector names should correspond to variable names}

\item{transformation}{Named list of \code{trans} objects. 
List names should correspond to variable names.}

\item{spill}{spillover matrix. If NULL, uncompensated data is used both for gating and plotting.}

\item{scale_values}{logical. Scale values for each y variable}

\item{free_y_scale}{logical. Free y scale (use a different scale for each y variable)}

\item{max_scale}{maximum scale for absolute y values. Used only if \code{scale_values = TRUE}}

\item{facet_vars}{variables used to facet plots along the x axis}

\item{group_var}{group variable}

\item{expand_factor}{numeric value expanding y scale. Used only if \code{free_y_scale = FALSE}}

\item{stat_function}{Name of the summary statistics function used}

\item{show.legend}{logical. Should plot legend be displayed?}

\item{y_trans}{default \code{trans} object to be used if \code{transformation} is NULL.}

\item{strip.text.y.angle}{angle of y strip labels}

\item{theme_name}{ggplot2 theme name ('theme_gray' by default)}

\item{Rowv}{Cluster rows using hierarchical clustering?}

\item{Colv}{Cluster columns using hierarchical clustering?}

\item{type}{plot type}

\item{metadata}{data.frame with metadata.}
}
\value{
a ggplot object
}
\description{
Plot summary statistics from a gating set
}
