% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGatingHierarchy_module.R
\name{plotGatingHierarchy}
\alias{plotGatingHierarchy}
\title{plotGatingHierarchy}
\usage{
plotGatingHierarchy(input, output, session, rval,
  plot_params = reactiveValues())
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactiveValues object}

\item{plot_params}{A reactiveValues object with plot parameters}
}
\value{
A list of plots
}
\description{
A shiny Module to plot a gating hierarchy
}
