% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGatingSet_module.R
\name{plotGatingSetInput}
\alias{plotGatingSetInput}
\title{plotGatingSetInput}
\usage{
plotGatingSetInput(id, simple_plot = TRUE, auto_update = TRUE)
}
\arguments{
\item{id}{shiny id}

\item{simple_plot}{logical, disable a number of plot options}

\item{auto_update}{Should plot update be automatic? 
If FALSE, plot update is controlled by an action button.}
}
\description{
A shiny Module (ui function) to build plots from a gating set
}
