% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting2_module.R, R/plotting_module.R
\name{plotting2}
\alias{plotting2}
\alias{plottingUI}
\alias{plotting}
\title{ploting2 server function}
\usage{
plotting2(input, output, session, rval)

plottingUI(id)

plotting(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}

\item{id}{shiny id}
}
\value{
The updated reactiveValues object \code{rval}

The updated reactiveValues object \code{rval}
}
\description{
A shiny Module to build, display and save plots from a gating set
}
