% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_module.R
\name{Plotting}
\alias{Plotting}
\title{Plotting module server function}
\usage{
Plotting(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{reactivevalues object with the following elements :
\describe{
  \item{gating_set}{: a GatingSet object}
}}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Plotting module server function
}
