% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUI_module.R
\name{plottingUI}
\alias{plottingUI}
\alias{plotting}
\title{plottingUI and plotting}
\usage{
plottingUI(id)

plotting(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "flow_set", "parameters" and "gates_flowCore"
}
\description{
A shiny Module that deals with metadata
}
