% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_module.R
\name{plottingUI}
\alias{plottingUI}
\alias{plotting}
\title{plottingUI and plotting}
\usage{
plottingUI(id)

plotting(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
A shiny Module to build, display and save plots from a gating set
}
