% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read-sheet.R
\name{read_sheet}
\alias{read_sheet}
\title{read_sheet}
\usage{
read_sheet(x, id_column, start_row = 1, sheet = "sample_sheet", ext,
  header = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{path to a file, to be read}

\item{id_column}{all rows which have this column as blank are skipped. See details.}

\item{start_row}{supplied to read.xlsx}

\item{sheet}{supplied to read.xlsx}

\item{ext}{determined using file extention. Specifying will override}

\item{header}{first line is header?}

\item{verbose}{be chatty?}

\item{...}{passed onto read.xlsx of openxlsx, read.table of read.csv2 depending on the file type.}
}
\description{
Auto detect file type, read it and clean it.
}
\details{
If id_column is missing, default if first column
}

