% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample-sheet.R
\name{read_sheet}
\alias{read_sheet}
\title{read_sheet}
\usage{
read_sheet(x, id_column, start_row = 1, sheet = "sample_sheet", ...)
}
\arguments{
\item{x}{}

\item{id_column}{all rows which have this column as blank are skipped. See details.}

\item{...}{passed onto \link{read.xlsx}, read.table of read.csv2 depending on the file type.}
}
\description{
Auto detect file type, read it and clean it.
}
\details{
If id_column is skipped the first column takes its place.
}

