% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rerun.R
\name{rerun}
\alias{rerun}
\alias{rerun.character}
\alias{rerun.flow}
\title{rerun}
\usage{
rerun(x, ...)

\method{rerun}{character}(x, ...)

\method{rerun}{flow}(x, mat, def, start_from, execute = TRUE, kill = TRUE,
  ...)
}
\arguments{
\item{x}{Either path to flow folder or the \link{flow} object which has been 'returned' from \link{submit_flow}.}

\item{...}{not used}

\item{mat}{path to flow_mat. should fetch on the fly}

\item{def}{path to should fetch on the fly}

\item{start_from}{which job to start from}

\item{execute}{[logical] whether to execute or not}

\item{kill}{logical indicating whether to kill the jobs from old flow}
}
\description{
rerun
}
\details{
We need path to the flow folder (\code{wd}). The \link{flow} object needs to have upate 'base_path' slow with wd (the path to the flow folder). Also its important to know that we need details regarding the previous submission from flow_details.txt file. Which should typically be in \code{wd}
}
\examples{
\dontrun{
rerun_flow(wd = wd, fobj = fobj, execute = TRUE, kill = TRUE)
}
}

