% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rerun-funcs.R
\name{rerun_flow}
\alias{rerun_flow}
\title{rerun_flow}
\usage{
rerun_flow(x, execute = TRUE, kill = TRUE)
}
\arguments{
\item{x}{Either path to flow folder or the \link{flow} object which has been 'returned' from \link{submit_flow}.}

\item{execute}{[logical] whether to execute or not}

\item{kill}{logical indicating whether to kill the jobs from old flow}
}
\description{
rerun_flow
}
\details{
We need path to the flow folder (\code{wd}). The \link{flow} object needs to have upate 'base_path' slow with wd (the path to the flow folder). Also its important to know that we need details regarding the previous submission from flow_details.txt file. Which should typically be in \code{wd}
}
\examples{
\dontrun{
rerun_flow(wd = wd, fobj = fobj, execute = TRUE, kill = TRUE)
}
}

