% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/examples.R
\name{run}
\alias{run}
\alias{run_flow}
\title{run pipelines
Running examples flows
This wraps a few steps:
Get all the commands to run (flow_mat)
Create a `flow` object, using flow_mat and a default flow_def (picked from the same folder).
Use `submit_flow()` to submit this to the cluster.}
\usage{
run(x = "sleep", type = "example", platform, flow_mat, flow_def,
  execute = FALSE, ...)
}
\arguments{
\item{x}{name of the pipeline to run. This is a function called to create a flow_mat.}

\item{type}{adv.}

\item{platform}{what platform to use, overrides flow_def}

\item{flow_mat}{flow matrix, with commands to run (if we already have this, start here)}

\item{flow_def}{flow definition}

\item{execute}{TRUE/FALSE}

\item{...}{passed onto the function used to create the flow_mat}
}
\description{
run pipelines
Running examples flows
This wraps a few steps:
Get all the commands to run (flow_mat)
Create a `flow` object, using flow_mat and a default flow_def (picked from the same folder).
Use `submit_flow()` to submit this to the cluster.
}

