% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{run_flow}
\alias{run_pipe}
\title{run pipelines}
\usage{
run(x, platform, def, flow_run_path = get_opts("flow_run_path"),
  execute = FALSE, ...)

run_pipe(x, platform, def, flow_run_path = get_opts("flow_run_path"),
  execute = FALSE, ...)
}
\arguments{
\item{x}{name of the pipeline to run. This is a function called to create a flow_mat.}

\item{platform}{what platform to use, overrides flowdef}

\item{def}{flow definition}

\item{flow_run_path}{passed onto to_flow. Default it picked up from flowr.conf. Typically this is ~/flowr/runs}

\item{execute}{TRUE/FALSE}

\item{...}{passed onto the pipeline function specified in x}
}
\description{
Running examples flows
This wraps a few steps:
Get all the commands to run (flow_mat)
Create a `flow` object, using flow_mat and a default flowdef (picked from the same folder).
Use `submit_flow()` to submit this to the cluster.
}

