% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow-def.R
\name{sample_flow_def}
\alias{sample_flow_def}
\title{Create a skeleton flow definition}
\usage{
sample_flow_def(fl, jobnames, cols)
}
\arguments{
\item{fl}{path to a matrix with commands to run}

\item{jobnames}{}
}
\description{
Create a skeleton flow definition
}
\details{
flow_tab: as defined by fl is a (minimum) three column matrix with
samplename, jobname, cmd
}

