% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariablesInput_module.R,
%   R/selectionInput_module.R
\name{selectionInput}
\alias{selectionInput}
\alias{selection}
\title{selectionInput and selection}
\usage{
selectionInput(id, multiple_subset = TRUE)

selection(input, output, session, rval, params = reactiveValues())

selectionInput(id, multiple_subset = TRUE)

selection(input, output, session, rval, params = reactiveValues())
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"

a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
A shiny Module that deals with metadata

A shiny Module that deals with metadata
}
