% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup and initialize flowr}
\usage{
setup(
  bin = "~/bin",
  flow_base_path = opts_flow$get("flow_base_path"),
  flow_run_path = opts_flow$get("flow_run_path"),
  flow_conf_path = opts_flow$get("flow_conf_path"),
  flow_pipe_path = opts_flow$get("flow_pipe_paths")
)
}
\arguments{
\item{bin}{path to bin folder}

\item{flow_base_path}{The base of flowr configuration and execution folders.}

\item{flow_run_path}{base path to be used for execution of this flow. 
flowr would create a new time-stamped folder in this base path and 
use it for logs, scripts etc. 
The default is retrieved using \code{opts_flow$get("flow_run_path")}.}

\item{flow_conf_path}{Flowr configuration folder, used by \link{fetch_conf}.}

\item{flow_pipe_path}{Folder with all pipelines, used by \link{fetch_pipes}.}
}
\description{
This functions creates a directory structure in user's home directory. 
Additionally it creates a shortcut to the \code{flowr} helper script in \code{~/bin}.
}
\details{
Will add more to this, to identify cluster and aid in other things.
}
