% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup and initialize some scripts.}
\usage{
setup(bin = "~/bin")
}
\arguments{
\item{bin}{path to bin folder}
}
\description{
Setup and initialize some scripts.
}
\details{
Will add more to this to identify cluster and aid in other things
}

