% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleDisplayUI_module.R
\name{simpleDisplayUI}
\alias{simpleDisplayUI}
\alias{simpleDisplay}
\title{simpleDisplayUI and simpleDisplay}
\usage{
simpleDisplayUI(id, nrow = 1, size = 400, save = TRUE)

simpleDisplay(input, output, session, plist, gate = reactiveValues(),
  params = reactiveValues(use_plotly = FALSE))
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "df_files", "flow_set_imported" and "gates_flowCore"
}
\description{
A shiny Module that deals with metadata
}
