% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_module.R
\name{StatsUI}
\alias{StatsUI}
\alias{Stats}
\title{Compute and visualize sample/subset statistics from a GatingSet}
\usage{
StatsUI(id)

Stats(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Compute and visualize sample/subset statistics from a GatingSet

Stats module server function
}
