% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{submit_flow,flow-method}
\alias{submit_flow}
\alias{submit_flow,flow-method}
\title{submit_flow}
\usage{
\S4method{submit_flow}{flow}(f_obj, attach_uuid = TRUE, execute = FALSE,
  make_flow_plot = TRUE, ...)
}
\arguments{
\item{f_obj}{\code{object} of class \code{flow}.}

\item{attach_uuid}{\code{logical} Whether to attach a random string to the submitted flow names. Look for them in your \code{basepath} folder (typically \code{~/flows/FLOW_NAME/FLOW_DESCRIPTION_UUID}). Refer to \code{desc} and \code{name} paramters of \link{flow}.}

\item{execute}{\code{logical} whether or not to submit the jobs}

\item{make_flow_plot}{\code{logical} whether to make a flow plot (saves it in the flow working directory)}

\item{...}{Any additional parameter are passed on to \link{submit_job} function}
}
\description{
submit_flow
}
\examples{
\dontrun{
submit_flow(f_obj = f_obj, ... = ...)}
}

