% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-funcs.R
\docType{methods}
\name{submit_flow,flow-method}
\alias{submit_flow}
\alias{submit_flow,flow-method}
\title{submit_flow}
\usage{
\S4method{submit_flow}{flow}(f_obj, uuid, execute = FALSE,
  make_flow_plot = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{f_obj}{\code{object} of class \code{flow}.}

\item{uuid}{\code{character} A character string pointing to the folder (unique) where all the logs and other files are processed. This is optional and defaults to:
\code{FLOW_DESCRIPTION_UUID}, and this folder is typically created in \code{~/flows/FLOW_NAME}.
Refer to \code{desc} and \code{name} paramters of \link{flow}.}

\item{execute}{\code{logical} whether or not to submit the jobs}

\item{make_flow_plot}{\code{logical} whether to make a flow plot (saves it in the flow working directory)}

\item{verbose}{logical.}

\item{...}{Any additional parameter are passed on to \link{submit_job} function}
}
\description{
submit_flow
}
\examples{
\dontrun{
submit_flow(f_obj = f_obj, ... = ...)}
}

