% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submit_flow.R
\name{submit_flow}
\alias{submit_flow}
\title{submit_flow}
\usage{
submit_flow(x, uuid, execute = FALSE, plot = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{a \code{object} of class \code{flow}.}

\item{uuid}{\code{character} Advanced use. Incase of rerunning the flow. uuid: typically is the final path of a previous flow.}

\item{execute}{\code{logical} whether or not to submit the jobs}

\item{plot}{\code{logical} whether to make a pdf flow plot (saves it in the flow working directory).}

\item{verbose}{logical.}

\item{...}{Advanced use. Any additional parameters are passed on to \link{submit_job} function.}
}
\description{
submit_flow
}
\examples{
\dontrun{
submit_flow(fobj = fobj, ... = ...)}
}

