% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class-funcs.R
\docType{methods}
\name{submit_job}
\alias{submit_job}
\alias{submit_job,job,flow,job-method}
\alias{submit_job,job,flow-method}
\title{submit_job}
\usage{
submit_job(j_obj, f_obj, ...)

\S4method{submit_job}{job,flow}(j_obj, f_obj, execute = FALSE,
  verbose = FALSE, wd, job_id, ...)
}
\arguments{
\item{j_obj}{Object of calls \link{job}}

\item{f_obj}{Object of calls \link{flow}}

\item{...}{not used}

\item{execute}{A \code{logical} vector suggesting whether to submit this job}

\item{verbose}{logical}

\item{wd}{working direcotry}

\item{job_id}{job id}
}
\description{
submit_job
}
\examples{
\dontrun{
submit_job(j_obj = j_obj, f_obj = f_obj, execute = FALSE,
verbose = TRUE, wd = wd, job_id = job_id)
}
}

