% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submit-job.R
\name{submit_job}
\alias{submit_job}
\title{submit_job}
\usage{
submit_job(jobj, fobj, job_id, execute = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{jobj}{Object of calls \link{job}}

\item{fobj}{Object of calls \link{flow}}

\item{job_id}{job id}

\item{execute}{A \code{logical} vector suggesting whether to submit this job}

\item{verbose}{logical}

\item{...}{not used}
}
\description{
submit_job
}
\examples{
\dontrun{
submit_job(jobj = jobj, fobj = fobj, execute = FALSE,
verbose = TRUE, wd = wd, job_id = job_id)
}
}

