% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsample_module.R
\name{SubsampleUI}
\alias{SubsampleUI}
\alias{Subsample}
\title{Subsample a GatingSet}
\usage{
SubsampleUI(id)

Subsample(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{reactivevalues object with the following elements :
\describe{
  \item{gating_set}{: a GatingSet object}
}}
}
\value{
The input reactivevalues object 'rval' with updated elements :
\describe{
  \item{gating_set_list}{list of GatingSet objects loaded}
  \item{gating_set}{selected GatingSet}
  \item{gating_set_selected}{Name of the selected GatingSet}
}
}
\description{
Subsample a GatingSet

Subsample module server function
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)
library(flowWorkspace)
library(flowCore)

if (interactive()){
  
  ui <- dashboardPage(
    dashboardHeader(title = "Subsample"),
    sidebar = dashboardSidebar(disable = TRUE),
    body = dashboardBody(
      fluidRow(
        column(12, box(width = NULL, SubsampleUI("module")))
      )
    )
  )
  
  server <- function(input, output, session) {
    
    rval <- reactiveValues()
    
    observe({
      utils::data("GvHD", package = "flowCore")
      rval$gating_set <- GatingSet(GvHD)
    })
    
    rval <- callModule(Subsample, "module", rval = rval)
    
  }
  
  shinyApp(ui, server)
  
}}
}
