% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rerun.R
\name{subset_fdef}
\alias{subset_fdef}
\title{subset_fdef}
\usage{
subset_fdef(fobj, def, start_from)
}
\arguments{
\item{fobj}{flow object}

\item{def}{flowdef}

\item{start_from}{where to start from}
}
\description{
subset_fdef
}

