% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rerun.R
\name{subset_fdet}
\alias{subset_fdet}
\title{subset flow details file}
\usage{
subset_fdet(fobj, det, start_from)
}
\arguments{
\item{fobj}{flow object}

\item{det}{flowdet}

\item{start_from}{where to start from}
}
\description{
subset flow details file
}

