% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rerun.R
\name{subset_fmat}
\alias{subset_fmat}
\title{subset_fmat}
\usage{
subset_fmat(fobj, mat, start_from)
}
\arguments{
\item{fobj}{flow object}

\item{mat}{a part of flowdef}

\item{start_from,}{where to start from}
}
\description{
subset_fmat
}

