% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flow.R
\name{to_flow}
\alias{to_flow}
\title{this operates on a single sample basis}
\usage{
to_flow(x, ...)
}
\arguments{
\item{x}{is a data.frame with jobnames and commands to run. See details for more on the format}
}
\description{
this operates on a single sample basis
}
\details{
subset the data.frame by sample and then supply to this function, if you want seperate flow for each sample.
flow_tab: as defined by x is a (minimum) three column matrix with:
samplename, jobname, cmd
}

