% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flowdef.R
\name{to_flowdef}
\alias{to_flowdef}
\alias{to_flowdef.character}
\alias{to_flowdef.flow}
\alias{to_flowdef.flowmat}
\title{Create a skeleton flow definition using a flowmat.}
\usage{
to_flowdef(x, ...)

\method{to_flowdef}{flowmat}(x, sub_type, dep_type, prev_jobs,
  queue = "short", platform = "torque", memory_reserved = "2000",
  cpu_reserved = "1", walltime = "1:00", ...)

\method{to_flowdef}{flow}(x, ...)

\method{to_flowdef}{character}(x, ...)
}
\arguments{
\item{x}{can a path to a flowmat, flomat or flow object.}

\item{...}{not used}

\item{sub_type}{submission type, one of: scatter, serial. Character, of length one or same as the number of jobnames}

\item{dep_type}{dependency type, one of: gather, serial or burst. Character, of length one or same as the number of jobnames}

\item{prev_jobs}{previous job name}

\item{queue}{Cluster queue to be used}

\item{platform}{platform of the cluster: lsf, sge, moab, torque, slurm etc.}

\item{memory_reserved}{amount of memory required.}

\item{cpu_reserved}{number of cpu's required}

\item{walltime}{amount of walltime required}
}
\description{
Creation of a skeleton flow definition with several default values.

All params may be of length one, or same as the number of jobnames

to_flowdef.character: x is a flowmat file.
}

