% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-flowmat.R
\name{to_flowmat}
\alias{to_flowmat}
\alias{to_flowmat.data.frame}
\alias{to_flowmat.flow}
\alias{to_flowmat.list}
\title{Taking in a named list and returns a two columns data.frame}
\usage{
to_flowmat(x, ...)

\method{to_flowmat}{list}(x, samplename, ...)

\method{to_flowmat}{data.frame}(x, ...)

\method{to_flowmat}{flow}(x, ...)
}
\arguments{
\item{x}{a named list OR vector. Where name corresponds to the jobname and value is a vector of commands to run}

\item{...}{not used}

\item{samplename}{character of length 1 or that of nrow(x)}
}
\description{
Taking in a named list and returns a two columns data.frame
}

