% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation_module.R, R/transform_module.R
\name{compensation}
\alias{compensation}
\alias{transformUI}
\alias{transform}
\title{compensation server function}
\usage{
compensation(input, output, session, rval)

transformUI(id)

transform(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}

\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}

The updated reactiveValues object \code{rval}
}
\description{
A shiny Module that deals with data transformation and parameters description
}
