% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_module.R
\name{TransformUI}
\alias{TransformUI}
\alias{Transform}
\title{Edit flow parameter transformation and description}
\usage{
TransformUI(id)

Transform(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
Edit flow parameter transformation and description

Transform module server function
}
