% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_module.R
\name{transformUI}
\alias{transformUI}
\alias{transform}
\title{transformUI and transform}
\usage{
transformUI(id)

transform(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
A shiny Module that deals with data transformation and parameters description
}
