% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensationUI_module.R, R/transformUI_module.R
\name{compensation}
\alias{compensation}
\alias{transformUI}
\alias{transform}
\title{compensation server function}
\usage{
compensation(input, output, session, rval)

transformUI(id)

transform(input, output, session, rval)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
a reactivevalues object with values "flow_set", "parameters" and "gates_flowCore"

a reactivevalues object with values "flow_set", "parameters" and "gates_flowCore"
}
\description{
A shiny Module that deals with metadata
}
