% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submit-job.R
\name{whisker_render}
\alias{whisker_render}
\title{Wrapper around whisker.render with some sugar on it...}
\usage{
whisker_render(template, data)
}
\arguments{
\item{template}{template used}

\item{data}{a list with variables to be used to fill in the template.}
}
\description{
This is a wrapper around \link{whisker.render}
}

