% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rw-flow-details.R
\name{write_flow_details}
\alias{write_flow_details}
\title{write files desribing this flow}
\usage{
write_flow_details(x, fobj, summ, flow_det, plot = FALSE)
}
\arguments{
\item{x}{path to write to}

\item{fobj}{flow object}

\item{summ}{a status summary.}

\item{flow_det}{a flow details data.frame}

\item{plot}{logical, plot or not}
}
\description{
write files desribing this flow
}

