% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw-flow-details.R
\name{write_flow_details}
\alias{write_flow_details}
\title{Write files describing this flow}
\usage{
write_flow_details(x, fobj, summ, flow_det, flow_mat, flow_def, plot = FALSE)
}
\arguments{
\item{x}{path to write to}

\item{fobj}{flow object}

\item{summ}{a status summary.}

\item{flow_det}{a flow details data.frame}

\item{flow_mat}{flow matrix (of commands)}

\item{flow_def}{flow definiion}

\item{plot}{logical, plot or not}
}
\description{
Write files describing this flow
}
