% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith_module.R
\name{zenithUI}
\alias{zenithUI}
\alias{zenith}
\title{zenithUI and zenith}
\usage{
zenithUI(id)

zenith(input, output, session, rval)
}
\arguments{
\item{id}{shiny id}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rval}{A reactive values object}
}
\value{
The updated reactiveValues object \code{rval}
}
\description{
A shiny Module that analyse cell metabolism using the ZeNITH method
}
